Extract.listof {spatstat.geom}	R Documentation

Extract or Replace Subset of a List of Things

Description

Replace a subset of a list of things.

Usage

  ## S3 replacement method for class 'listof'
x[i] <- value

Arguments

x
An object of class "listof" representing a list of things which all belong to one class.

i
Subset index. Any valid subset index in the usual R sense.

value
Replacement value for the subset.

Details

This is a subset replacement method for the class "listof".

The argument x should be an object of class "listof" representing a list of things that all belong to one class.

The method replaces a designated subset of x, and returns an object of class "listof".

Value

Another object of class "listof".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

plot.listof, summary.listof

Examples

   x <- list(A=runif(10), B=runif(10), C=runif(10))
   class(x) <- c("listof", class(x))
   x[1] <- list(A=rnorm(10))
 

[Package spatstat.geom version 3.3-5 Index]