Extract.owin {spatstat.geom}	R Documentation

Extract Subset of Window

Description

Extract a subset of a window.

Usage

  ## S3 method for class 'owin'
x[i, ...]

Arguments

x
A spatial window (object of class "owin").

i
Object defining the subregion. Either a spatial window, or a pixel image with logical values.

...
Ignored.

Details

This function computes the intersection between the window x and the domain specified by i, using intersect.owin.

This function is a method for the subset operator "[" for spatial windows (objects of class "owin"). It is provided mainly for completeness.

The index i may be either a window, or a pixel image with logical values (the TRUE values of the image specify the spatial domain).

Value

Another spatial window (object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

intersect.owin

Examples

 W <- owin(c(2.5, 3.2), c(1.4, 2.9))
 plot(letterR)
 plot(letterR[W], add=TRUE, col="red")

[Package spatstat.geom version 3.3-5 Index]