Extract.quad {spatstat.geom}	R Documentation

Subset of Quadrature Scheme

Description

Extract a subset of a quadrature scheme.

Usage

  ## S3 method for class 'quad'
x[...]

Arguments

x
A quadrature scheme (object of class "quad").

...
Arguments passed to [.ppp to determine the subset.

Details

This function extracts a designated subset of a quadrature scheme.

The function [.quad is a method for [ for the class "quad". It extracts a designated subset of a quadrature scheme.

The subset to be extracted is determined by the arguments ... which are interpreted by [.ppp. Thus it is possible to take the subset consisting of all quadrature points that lie inside a given region, or a subset of quadrature points identified by numeric indices.

Value

A quadrature scheme (object of class "quad").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

quad.object, [.ppp.

Examples

 Q <- quadscheme(nztrees)
 W <- owin(c(0,148),c(0,95)) # a subwindow
 Q[W]

[Package spatstat.geom version 3.3-5 Index]