Extract.solist {spatstat.geom}	R Documentation

Extract or Replace Subset of a List of Spatial Objects

Description

Extract or replace some entries in a list of spatial objects, or extract a designated sub-region in each object.

Usage

  ## S3 method for class 'solist'
x[i, ...]

  ## S3 replacement method for class 'solist'
x[i] <- value

Arguments

x
An object of class "solist" representing a list of two-dimensional spatial objects.

i
Subset index. Any valid subset index for vectors in the usual R sense, or a window (object of class "owin").

value
Replacement value for the subset.

...
Ignored.

Details

These are methods for extracting and replacing subsets for the class "solist".

The argument x should be an object of class "solist" representing a list of two-dimensional spatial objects. See solist.

For the subset method, the subset index i can be either a vector index (specifying some elements of the list) or a spatial window (specifying a spatial sub-region).

For the replacement method, i must be a vector index: the designated elements will be replaced.

Value

Another object of the same class as x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

solist, plot.solist, summary.solist

Examples

   x <- solist(japanesepines, cells, redwood)
   x[2:3]
   x[square(0.5)]
   x[1] <- list(finpines)
 

[Package spatstat.geom version 3.3-5 Index]