Extract.splitppp {spatstat.geom}	R Documentation

Extract or Replace Sub-Patterns

Description

Extract or replace some of the sub-patterns in a split point pattern.

Usage

  ## S3 method for class 'splitppp'
x[...]
  ## S3 replacement method for class 'splitppp'
x[...] <- value

Arguments

x
An object of class "splitppp", representing a point pattern separated into a list of sub-patterns.

...
Subset index. Any valid subset index in the usual R sense.

value
Replacement value for the subset. A list of point patterns.

Details

These are subset methods for the class "splitppp".

The argument x should be an object of class "splitppp", representing a point pattern that has been separated into a list of sub-patterns. It is created by split.ppp.

The methods extract or replace a designated subset of the list x, and return an object of class "splitppp".

Value

Another object of class "splitppp".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

split.ppp, plot.splitppp, summary.splitppp

Examples

  y <- split(amacrine)
  y[[1]]
  y[["off"]]
  y[[1]] <- rsyst(Window(amacrine), 4, 3)
 

[Package spatstat.geom version 3.3-5 Index]