Frame {spatstat.geom}	R Documentation

Extract or Change the Containing Rectangle of a Spatial Object

Description

Given a spatial object (such as a point pattern or pixel image) in two dimensions, these functions extract or change the containing rectangle inside which the object is defined.

Usage

   Frame(X)

   ## Default S3 method:
Frame(X)

   Frame(X) <- value

   ## S3 replacement method for class 'owin'
Frame(X) <- value

   ## S3 replacement method for class 'ppp'
Frame(X) <- value

   ## S3 replacement method for class 'im'
Frame(X) <- value

   ## Default S3 replacement method:
Frame(X) <- value

Arguments

X
A spatial object such as a point pattern, line segment pattern or pixel image.

value
A rectangular window (object of class "owin" of type "rectangle") to be used as the new containing rectangle for X.

Details

The functions Frame and Frame<- are generic.

Frame(X) extracts the rectangle inside which X is defined.

Frame(X) <- R changes the rectangle inside which X is defined to the new rectangle R.

Value

The result of Frame is a rectangular window (object of class "owin" of type "rectangle").

The result of Frame<- is the updated object X, of the same class as X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

Window

Examples

   Frame(cells)
   X <- demopat
   Frame(X)
   Frame(X) <- owin(c(0, 11000), c(400, 8000))

[Package spatstat.geom version 3.3-5 Index]