Window.tess {spatstat.geom}	R Documentation

Extract Window of Spatial Object

Description

Given a spatial object (such as a point pattern or pixel image) in two dimensions, these functions extract the window in which the object is defined.

Usage

 ## S3 method for class 'quadratcount'
Window(X, ...)

 ## S3 method for class 'tess'
Window(X, ...)

 ## S3 method for class 'layered'
Window(X, ...)

 ## S3 method for class 'distfun'
Window(X, ...)

 ## S3 method for class 'nnfun'
Window(X, ...)

 ## S3 method for class 'funxy'
Window(X, ...)

Arguments

X
A spatial object.

...
Ignored.

Details

These are methods for the generic function Window which extract the spatial window in which the object X is defined.

Value

An object of class "owin" (see owin.object) specifying an observation window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

Window, Window.ppp, Window.psp.

owin.object

Examples

   A <- quadratcount(cells, 4)
   Window(A)

[Package spatstat.geom version 3.3-5 Index]