add.texture {spatstat.geom}	R Documentation

Fill Plot With Texture

Description

Draws a simple texture inside a region on the plot.

Usage

add.texture(W, texture = 4, spacing = NULL, ...)

Arguments

W
Window (object of class "owin") inside which the texture should be drawn.

texture
Integer from 1 to 8 identifying the type of texture. See Details.

spacing
Spacing between elements of the texture, in units of the current plot.

...
Further arguments controlling the plot colour, line width etc.

Details

The chosen texture, confined to the window W, will be added to the current plot. The available textures are:

texture=1:

Small crosses arranged in a square grid.

texture=2:

Parallel vertical lines.

texture=3:

Parallel horizontal lines.

texture=4:

Parallel diagonal lines at 45 degrees from the horizontal.

texture=5:

Parallel diagonal lines at 135 degrees from the horizontal.

texture=6:

Grid of horizontal and vertical lines.

texture=7:

Grid of diagonal lines at 45 and 135 degrees from the horizontal.

texture=8:

Grid of hexagons.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

owin, plot.owin, textureplot, texturemap.

Examples

  W <- Window(chorley)
  plot(W, main="")
  add.texture(W, 7)

[Package spatstat.geom version 3.3-5 Index]