affine {spatstat.geom}	R Documentation

Apply Affine Transformation

Description

Applies any affine transformation of the plane (linear transformation plus vector shift) to a plane geometrical object, such as a point pattern or a window.

Usage

  affine(X, ...)

Arguments

X
Any suitable dataset representing a two-dimensional object, such as a point pattern (object of class "ppp"), a line segment pattern (object of class "psp"), a window (object of class "owin") or a pixel image (object of class "im").

...
Arguments determining the affine transformation.

Details

This is generic. Methods are provided for point patterns (affine.ppp) and windows (affine.owin).

Value

Another object of the same type, representing the result of applying the affine transformation.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

affine.ppp, affine.psp, affine.owin, affine.im, flipxy, reflect, rotate, shift

[Package spatstat.geom version 3.3-5 Index]