anyNA.im {spatstat.geom}	R Documentation

Check Whether Image Contains NA Values

Description

Checks whether any pixel values in a pixel image are NA (meaning that the pixel lies outside the domain of definition of the image).

Usage

  ## S3 method for class 'im'
anyNA(x, recursive = FALSE)

Arguments

x
A pixel image (object of class "im").

recursive
Ignored.

Details

The function anyNA is generic: anyNA(x) is a faster alternative to any(is.na(x)).

This function anyNA.im is a method for the generic anyNA defined for pixel images. It returns the value TRUE if any of the pixel values in x are NA, and and otherwise returns FALSE.

Value

A single logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

im.object

Examples

  anyNA(as.im(letterR))

[Package spatstat.geom version 3.3-5 Index]