anylist {spatstat.geom}	R Documentation

List of Objects

Description

Make a list of objects of any type.

Usage

anylist(...)
as.anylist(x)

Arguments

...
Any number of arguments of any type.

x
A list.

Details

An object of class "anylist" is a list of objects that the user intends to treat in a similar fashion.

For example it may be desired to plot each of the objects side-by-side: this can be done using the function plot.anylist.

The objects can belong to any class; they may or may not all belong to the same class.

In the spatstat package, various functions produce an object of class "anylist".

Value

A list, belonging to the class "anylist", containing the original objects.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

solist, as.solist, anylapply.

Examples

  if(require(spatstat.explore)) {
    anylist(cells, intensity(cells), Kest(cells))
  } else {
    anylist(cells, intensity(cells))
  }
  anylist()

[Package spatstat.geom version 3.3-5 Index]