append.psp {spatstat.geom}	R Documentation

Combine Two Line Segment Patterns

Description

Combine two line segment patterns into a single pattern.

Usage

  append.psp(A, B)

Arguments

A,B
Line segment patterns (objects of class "psp").

Details

This function is used to superimpose two line segment patterns A and B.

The two patterns must have identical windows. If one pattern has marks, then the other must also have marks of the same type. It the marks are data frames then the number of columns of these data frames, and the names of the columns must be identical.

(To combine two point patterns, see superimpose).

If one of the arguments is NULL, it will be ignored and the other argument will be returned.

Value

Another line segment pattern (object of class "psp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

psp, as.psp, superimpose,

Examples

  X <- psp(runif(20), runif(20), runif(20), runif(20),  window=owin())
  Y <- psp(runif(5), runif(5), runif(5), runif(5),  window=owin())
  append.psp(X,Y)

[Package spatstat.geom version 3.3-5 Index]