areaLoss {spatstat.geom}	R Documentation

Difference of Disc Areas

Description

Computes the area of that part of a disc that is not covered by other discs.

Usage

   areaLoss(X, r, ..., W=as.owin(X), subset=NULL,
                 exact=FALSE,
                 ngrid=spatstat.options("ngrid.disc"))

Arguments

X
Locations of the centres of discs. A point pattern (object of class "ppp").

r
Disc radius, or vector of disc radii.

...
Ignored.

W
Optional. Window (object of class "owin") inside which the area should be calculated.

subset
Optional. Index identifying a subset of the points of X for which the area difference should be computed.

exact
Choice of algorithm. If exact=TRUE, areas are computed exactly using analytic geometry. If exact=FALSE then a faster algorithm is used to compute a discrete approximation to the areas.

ngrid
Integer. Number of points in the square grid used to compute the discrete approximation, when exact=FALSE.

Details

This function computes, for each point X[i] in X and for each radius r, the area of that part of the disc of radius r centred at the location X[i] that is not covered by any of the other discs of radius r centred at the points X[j] for j not equal to i. This area is important in some calculations related to the area-interaction model AreaInter.

The result is a matrix, with one row for each point in X and one column for each entry of r.

Value

A matrix with one row for each point in X (or X[subset]) and one column for each value in r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

AreaInter, areaGain, dilated.areas

Examples

   areaLoss(cells, 0.1)

[Package spatstat.geom version 3.3-5 Index]