as.box3 {spatstat.geom}	R Documentation

Convert Data to Three-Dimensional Box

Description

Interprets data as the dimensions of a three-dimensional box.

Usage

as.box3(...)

Arguments

...
Data that can be interpreted as giving the dimensions of a three-dimensional box. See Details.

Details

This function converts data in various formats to an object of class "box3" representing a three-dimensional box (see box3). The arguments ... may be

an object of class "box3"

arguments acceptable to box3

a numeric vector of length 6, interpreted as c(xrange[1],xrange[2],yrange[1],yrange[2],zrange[1],zrange[2])

an object of class "pp3" representing a three-dimensional point pattern contained in a box.

Value

Object of class "box3".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

box3, pp3

Examples

    X <- c(0,10,0,10,0,5)
    as.box3(X)
    X <- pp3(runif(42),runif(42),runif(42), box3(c(0,1)))
    as.box3(X)

[Package spatstat.geom version 3.3-5 Index]