as.boxx {spatstat.geom}	R Documentation

Convert Data to Multi-Dimensional Box

Description

Interprets data as the dimensions of a multi-dimensional box.

Usage

  as.boxx(..., warn.owin = TRUE)

Arguments

...
Data that can be interpreted as giving the dimensions of a multi-dimensional box. See Details.

warn.owin
Logical value indicating whether to print a warning if a non-rectangular window (object of class "owin") is supplied.

Details

Either a single argument should be provided which is one of the following:

an object of class "boxx"

an object of class "box3"

an object of class "owin"

a numeric vector of even length, specifying the corners of the box. See Examples

or a list of arguments acceptable to boxx.

Value

A "boxx" object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

Examples

 # Convert unit square to two dimensional box.
 W <- owin()
 as.boxx(W)
 # Make three dimensional box [0,1]x[0,1]x[0,1] from numeric vector
 as.boxx(c(0,1,0,1,0,1))

[Package spatstat.geom version 3.3-5 Index]