as.data.frame.hyperframe {spatstat.geom}	R Documentation

Coerce Hyperframe to Data Frame

Description

Converts a hyperframe to a data frame.

Usage

## S3 method for class 'hyperframe'
as.data.frame(x, row.names = NULL,
                                  optional = FALSE, ..., 
                                  discard=TRUE, warn=TRUE)

Arguments

x
Hyperframe (object of class "hyperframe").

row.names
Optional character vector of row names.

optional
Argument passed to as.data.frame controlling what happens to row names.

...
Ignored.

discard
Logical. Whether to discard columns of the hyperframe that do not contain atomic data. See Details.

warn
Logical. Whether to issue a warning when columns are discarded.

Details

This is a method for the generic function as.data.frame for the class of hyperframes (see hyperframe.

If discard=TRUE, any columns of the hyperframe that do not contain atomic data will be removed (and a warning will be issued if warn=TRUE). If discard=FALSE, then such columns are converted to strings indicating what class of data they originally contained.

Value

A data frame.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

Examples

  h <- hyperframe(X=1:3, Y=letters[1:3], f=list(sin, cos, tan))
  as.data.frame(h, discard=TRUE, warn=FALSE)
  as.data.frame(h, discard=FALSE)

[Package spatstat.geom version 3.3-5 Index]