as.data.frame.ppp {spatstat.geom}	R Documentation

Coerce Point Pattern to a Data Frame

Description

Extracts the coordinates of the points in a point pattern, and their marks if any, and returns them in a data frame.

Usage

## S3 method for class 'ppp'
as.data.frame(x, row.names = NULL, ...)

Arguments

x
Point pattern (object of class "ppp").

row.names
Optional character vector of row names.

...
Ignored.

Details

This is a method for the generic function as.data.frame for the class "ppp" of point patterns.

It extracts the coordinates of the points in the point pattern, and returns them as columns named x and y in a data frame. If the points were marked, the marks are returned as a column named marks with the same type as in the point pattern dataset.

Value

A data frame.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

Examples

  df <- as.data.frame(amacrine)
  df[1:5,]

[Package spatstat.geom version 3.3-5 Index]