as.data.frame.psp {spatstat.geom}	R Documentation

Coerce Line Segment Pattern to a Data Frame

Description

Extracts the coordinates of the endpoints in a line segment pattern, and their marks if any, and returns them in a data frame.

Usage

## S3 method for class 'psp'
as.data.frame(x, row.names = NULL, ...)

Arguments

x
Line segment pattern (object of class "psp").

row.names
Optional character vector of row names.

...
Ignored.

Details

This is a method for the generic function as.data.frame for the class "psp" of line segment patterns.

It extracts the coordinates of the endpoints of the line segments, and returns them as columns named x0, y0, x1 and y1 in a data frame. If the line segments were marked, the marks are appended as an extra column or columns to the data frame which is returned. If the marks are a vector then a single column named marks is appended. in the data frame, with the same type as in the line segment pattern dataset. If the marks are a data frame, then the columns of this data frame are appended (retaining their names).

Value

A data frame with 4 or 5 columns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

Examples

  df <- as.data.frame(copper$Lines)

[Package spatstat.geom version 3.3-5 Index]