as.data.frame.tess {spatstat.geom}	R Documentation

Convert Tessellation to Data Frame

Description

Converts a spatial tessellation object to a data frame.

Usage

## S3 method for class 'tess'
as.data.frame(x, ...)

Arguments

x
Tessellation (object of class "tess").

...
Further arguments passed to as.data.frame.owin or as.data.frame.im and ultimately to as.data.frame.default to determine the row names and other features.

Details

This function converts the tessellation x to a data frame.

If x is a pixel image tessellation (a pixel image with factor values specifying the tile membership of each pixel) then this pixel image is converted to a data frame by as.data.frame.im. The result is a data frame with columns x and y giving the pixel coordinates, and Tile identifying the tile containing the pixel.

If x is a tessellation consisting of a rectangular grid of tiles or a list of polygonal tiles, then each tile is converted to a data frame by as.data.frame.owin, and these data frames are joined together, yielding a single large data frame containing columns x, y giving the coordinates of vertices of the polygons, and Tile identifying the tile.

Value

A data frame with columns named x, y, Tile, and possibly other columns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

as.data.frame.owin, as.data.frame.im

Examples

  Z <- as.data.frame(dirichlet(cells))
  head(Z, 10)

[Package spatstat.geom version 3.3-5 Index]