as.layered {spatstat.geom}	R Documentation

Convert Data To Layered Object

Description

Converts spatial data into a layered object.

Usage

 as.layered(X)

 ## Default S3 method:
as.layered(X)

 ## S3 method for class 'ppp'
as.layered(X)

 ## S3 method for class 'splitppp'
as.layered(X)

 ## S3 method for class 'solist'
as.layered(X)

 ## S3 method for class 'listof'
as.layered(X)

Arguments

X
Some kind of spatial data.

Details

This function converts the object X into an object of class "layered".

The argument X should contain some kind of spatial data such as a point pattern, window, or pixel image.

If X is a simple object then it will be converted into a layered object containing only one layer which is equivalent to X.

If X can be interpreted as consisting of multiple layers of data, then the result will be a layered object consisting of these separate layers of data.

if X is a list of class "listof" or "solist", then as.layered(X) consists of several layers, one for each entry in the list X;

if X is a multitype point pattern, then as.layered(X) consists of several layers, each containing the sub-pattern consisting of points of one type;

if X is a vector-valued measure, then as.layered(X) consists of several layers, each containing a scalar-valued measure.

Value

An object of class "layered" (see layered).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

as.layered.msr, layered, split.ppp

Examples

   as.layered(cells)
   as.layered(amacrine)

[Package spatstat.geom version 3.3-5 Index]