bdist.points {spatstat.geom}	R Documentation

Distance to Boundary of Window

Description

Computes the distances from each point of a point pattern to the boundary of the window.

Usage

 bdist.points(X)

Arguments

X
A point pattern (object of class "ppp").

Details

This function computes, for each point x[i] in the point pattern X, the shortest distance dist(x[i], W') from x[i] to the boundary of the window W of observation.

If the window Window(X) is of type "rectangle" or "polygonal", then these distances are computed by analytic geometry and are exact, up to rounding errors. If the window is of type "mask" then the distances are computed using the real-valued distance transform, which is an approximation with maximum error equal to the width of one pixel in the mask.

Value

A numeric vector, giving the distances from each point of the pattern to the boundary of the window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

bdist.pixels, bdist.tiles, ppp.object, erosion

Examples

  d <- bdist.points(cells)

[Package spatstat.geom version 3.3-5 Index]