bdist.tiles {spatstat.geom}	R Documentation

Distance to Boundary of Window

Description

Computes the shortest distances from each tile in a tessellation to the boundary of the window.

Usage

 bdist.tiles(X)

Arguments

X
A tessellation (object of class "tess").

Details

This function computes, for each tile s[[i]] in the tessellation X, the shortest distance from s[[i]] to the boundary of the window W containing the tessellation.

Value

A numeric vector, giving the shortest distance from each tile in the tessellation to the boundary of the window. Entries of the vector correspond to the entries of tiles(X).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

tess, bdist.points, bdist.pixels

Examples

  P <- runifrect(15)
  X <- dirichlet(P)
  plot(X, col="red")
  B <- bdist.tiles(X)
  # identify tiles that do not touch the boundary
  plot(X[B > 0], add=TRUE, col="green", lwd=3)

[Package spatstat.geom version 3.3-5 Index]