box3 {spatstat.geom}	R Documentation

Three-Dimensional Box

Description

Creates an object representing a three-dimensional box.

Usage

box3(xrange = c(0, 1), yrange = xrange, zrange = yrange, unitname = NULL)

Arguments

xrange, yrange, zrange
Dimensions of the box in the x,y,z directions. Each of these arguments should be a numeric vector of length 2.

unitname
Optional. Name of the unit of length. See Details.

Details

This function creates an object representing a three-dimensional rectangular parallelepiped (box) with sides parallel to the coordinate axes.

The object can be used to specify the domain of a three-dimensional point pattern (see pp3) and in various geometrical calculations (see volume.box3, diameter.box3, eroded.volumes).

The optional argument unitname specifies the name of the unit of length. See unitname for valid formats.

The function as.box3 can be used to convert other kinds of data to this format.

Value

An object of class "box3". There is a print method for this class.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

as.box3, pp3, volume.box3, diameter.box3, eroded.volumes.

Examples

    box3()
    box3(c(0,10),c(0,10),c(0,5), unitname=c("metre","metres"))
    box3(c(-1,1))

[Package spatstat.geom version 3.3-5 Index]