boxx {spatstat.geom}	R Documentation

Multi-Dimensional Box

Description

Creates an object representing a multi-dimensional box.

Usage

boxx(..., unitname = NULL)

Arguments

...
Dimensions of the box. Vectors of length 2.

unitname
Optional. Name of the unit of length. See Details.

Details

This function creates an object representing a multi-dimensional rectangular parallelepiped (box) with sides parallel to the coordinate axes.

The object can be used to specify the domain of a multi-dimensional point pattern (see ppx) and in various geometrical calculations (see volume.boxx, diameter.boxx, eroded.volumes).

The optional argument unitname specifies the name of the unit of length. See unitname for valid formats.

Value

An object of class "boxx". There is a print method for this class.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

ppx, volume.boxx, diameter.boxx, eroded.volumes.boxx.

Examples

    boxx(c(0,10),c(0,10),c(0,5),c(0,1), unitname=c("metre","metres"))

[Package spatstat.geom version 3.3-5 Index]