chop.tess {spatstat.geom}	R Documentation

Subdivide a Window or Tessellation using a Set of Lines

Description

Divide a given window into tiles delineated by a set of infinite straight lines, obtaining a tessellation of the window. Alternatively, given a tessellation, divide each tile of the tessellation into sub-tiles delineated by the lines.

Usage

chop.tess(X, L)

Arguments

X
A window (object of class "owin") or tessellation (object of class "tess") to be subdivided by lines.

L
A set of infinite straight lines (object of class "infline")

Details

The argument L should be a set of infinite straight lines in the plane (stored in an object L of class "infline" created by the function infline).

If X is a window, then it is divided into tiles delineated by the lines in L.

If X is a tessellation, then each tile of X is subdivided into sub-tiles delineated by the lines in L.

The result is a tessellation.

Value

A tessellation (object of class "tess").

Warning

If X is a non-convex window, or a tessellation containing non-convex tiles, then chop.tess(X,L) may contain a tile which consists of several unconnected pieces.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

infline, clip.infline

Examples

  L <- infline(p=1:3, theta=pi/4)
  W <- square(4)
  chop.tess(W, L)

[Package spatstat.geom version 3.3-5 Index]