clickbox {spatstat.geom}	R Documentation

Interactively Define a Rectangle

Description

Allows the user to specify a rectangle by point-and-click in the display.

Usage

  clickbox(add=TRUE, ...)

Arguments

add
Logical value indicating whether to create a new plot (add=FALSE) or draw over the existing plot (add=TRUE).

...
Graphics arguments passed to polygon to plot the box.

Details

This function allows the user to create a rectangular window by interactively clicking on the screen display.

The user is prompted to point the mouse at any desired locations for two corners of the rectangle, and click the left mouse button to add each point.

The return value is a window (object of class "owin") representing the rectangle.

This function uses the R command locator to input the mouse clicks. It only works on screen devices such as ‘X11’, ‘windows’ and ‘quartz’.

Value

A window (object of class "owin") representing the selected rectangle.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

clickpoly, clickppp, clickdist, locator

[Package spatstat.geom version 3.3-5 Index]