clickdist {spatstat.geom}	R Documentation

Interactively Measure Distance

Description

Measures the distance between two points which the user has clicked on.

Usage

  clickdist()

Details

This function allows the user to measure the distance between two spatial locations, interactively, by clicking on the screen display.

When clickdist() is called, the user is expected to click two points in the current graphics device. The distance between these points will be returned.

This function uses the R command locator to input the mouse clicks. It only works on screen devices such as ‘X11’, ‘windows’ and ‘quartz’.

Value

A single nonnegative number.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

locator, clickppp, clickpoly, clickbox

[Package spatstat.geom version 3.3-5 Index]