clickpoly {spatstat.geom}	R Documentation

Interactively Define a Polygon

Description

Allows the user to create a polygon by point-and-click in the display.

Usage

  clickpoly(add=FALSE, nv=NULL, np=1, ...)

Arguments

add
Logical value indicating whether to create a new plot (add=FALSE) or draw over the existing plot (add=TRUE).

nv
Number of vertices of the polygon (if this is predetermined).

np
Number of polygons to create.

...
Arguments passed to locator to control the interactive plot, and to polygon to plot the polygons.

Details

This function allows the user to create a polygonal window by interactively clicking on the screen display.

The user is prompted to point the mouse at any desired locations for the polygon vertices, and click the left mouse button to add each point. Interactive input stops after nv clicks (if nv was given) or when the middle mouse button is pressed.

The return value is a window (object of class "owin") representing the polygon.

This function uses the R command locator to input the mouse clicks. It only works on screen devices such as ‘X11’, ‘windows’ and ‘quartz’. Arguments that can be passed to locator through ... include pch (plotting character), cex (character expansion factor) and col (colour). See locator and par.

Multiple polygons can also be drawn, by specifying np > 1. The polygons must be disjoint. The result is a single window object consisting of all the polygons.

Value

A window (object of class "owin") representing the polygon.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

identify.ppp, clickbox, clickppp, clickdist, locator

[Package spatstat.geom version 3.3-5 Index]