clip.infline {spatstat.geom}	R Documentation

Intersect Infinite Straight Lines with a Window

Description

Take the intersection between a set of infinite straight lines and a window, yielding a set of line segments.

Usage

clip.infline(L, win)

Arguments

L
Object of class "infline" specifying a set of infinite straight lines in the plane.

win
Window (object of class "owin").

Details

This function computes the intersection between a set of infinite straight lines in the plane (stored in an object L of class "infline" created by the function infline) and a window win. The result is a pattern of line segments. Each line segment carries a mark indicating which line it belongs to.

Value

A line segment pattern (object of class "psp") with a single column of marks.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

infline,psp.

To divide a window into pieces using infinite lines, use chop.tess.

Examples

  L <- infline(p=1:3, theta=pi/4)
  W <- square(4)
  clip.infline(L, W)

[Package spatstat.geom version 3.3-5 Index]