closepairs.pp3 {spatstat.geom}	R Documentation

Close Pairs of Points in 3 Dimensions

Description

Low-level functions to find all close pairs of points in three-dimensional point patterns.

Usage

## S3 method for class 'pp3'
closepairs(X, rmax, twice=TRUE,
                         what=c("all", "indices", "ijd"),
                         distinct=TRUE, neat=TRUE, ...)

## S3 method for class 'pp3'
crosspairs(X, Y, rmax, what=c("all", "indices", "ijd"), ...) 

Arguments

X,Y
Point patterns in three dimensions (objects of class "pp3").

rmax
Maximum distance between pairs of points to be counted as close pairs.

twice
Logical value indicating whether all ordered pairs of close points should be returned. If twice=TRUE, each pair will appear twice in the output, as (i,j) and again as (j,i). If twice=FALSE, then each pair will appear only once, as the pair (i,j) such that i < j.

what
String specifying the data to be returned for each close pair of points. If what="all" (the default) then the returned information includes the indices i,j of each pair, their x,y,z coordinates, and the distance between them. If what="indices" then only the indices i,j are returned. If what="ijd" then the indices i,j and the distance d are returned.

distinct
Logical value indicating whether to return only the pairs of points with different indices i and j (distinct=TRUE, the default) or to also include the pairs where i=j (distinct=FALSE).

neat
Logical value indicating whether to ensure that i < j in each output pair, when twice=FALSE.

...
Ignored.

Details

These are the efficient low-level functions used by spatstat to find all close pairs of points in a three-dimensional point pattern or all close pairs between two point patterns in three dimensions.

closepairs(X,rmax) identifies all pairs of neighbours in the pattern X and returns them. The result is a list with the following components:

i

Integer vector of indices of the first point in each pair.

j

Integer vector of indices of the second point in each pair.

xi,yi,zi

Coordinates of the first point in each pair.

xj,yj,zj

Coordinates of the second point in each pair.

dx

Equal to xj-xi

dy

Equal to yj-yi

dz

Equal to zj-zi

d

Euclidean distance between each pair of points.

If what="indices" then only the components i and j are returned. This is slightly faster.

crosspairs(X,rmax) identifies all pairs of neighbours (X[i], Y[j]) between the patterns X and Y, and returns them. The result is a list with the same format as for closepairs.

Value

A list with components i and j, and possibly other components as described under Details.

Warning about accuracy

The results of these functions may not agree exactly with the correct answer (as calculated by a human) and may not be consistent between different computers and different installations of R. The discrepancies arise in marginal cases where the interpoint distance is equal to, or very close to, the threshold rmax.

Floating-point numbers in a computer are not mathematical Real Numbers: they are approximations using finite-precision binary arithmetic. The approximation is accurate to a tolerance of about .Machine$double.eps.

If the true interpoint distance d and the threshold rmax are equal, or if their difference is no more than .Machine$double.eps, the result may be incorrect.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

, Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk.

See Also

closepairs

Examples

   X <- pp3(runif(10), runif(10), runif(10), box3(c(0,1)))
   Y <- pp3(runif(10), runif(10), runif(10), box3(c(0,1)))
   a <- closepairs(X, 0.1)
   b <- crosspairs(X, Y, 0.1)

[Package spatstat.geom version 3.3-5 Index]