closetriples {spatstat.geom}	R Documentation

Close Triples of Points

Description

Low-level function to find all close triples of points.

Usage

closetriples(X, rmax)

Arguments

X
Point pattern (object of class "ppp" or "pp3").

rmax
Maximum distance between each pair of points in a triple.

Details

This low-level function finds all triples of points in a point pattern in which each pair lies closer than rmax.

Value

A data frame with columns i,j,k giving the indices of the points in each triple, and a column diam giving the diameter (maximum pairwise distance) in the triple.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

closepairs, Tstat.

Examples

   closetriples(redwoodfull, 0.02)
   closetriples(redwoodfull, 0.005)

[Package spatstat.geom version 3.3-5 Index]