commonGrid {spatstat.geom}	R Documentation

Determine A Common Spatial Domain And Pixel Resolution

Description

Determine a common spatial domain and pixel resolution for several spatial objects such as images, masks, windows and point patterns.

Usage

commonGrid(...)

Arguments

...
Any number of pixel images (objects of class "im"), binary masks (objects of class "owin" of type "mask") or data which can be converted to binary masks by as.mask.

Details

This function determines a common spatial resolution and spatial domain for several spatial objects.

The arguments ... may be pixel images, binary masks, or other spatial objects acceptable to as.mask.

The common pixel grid is determined by inspecting all the pixel images and binary masks in the argument list, finding the pixel grid with the highest spatial resolution, and extending this pixel grid to cover the bounding box of all the spatial objects.

The return value is a binary mask M, representing the bounding box at the chosen pixel resolution. Use as.im(X, W=M) to convert a pixel image X to this new pixel resolution. Use as.mask(W, xy=M) to convert a window W to a binary mask at this new pixel resolution. See the Examples.

Value

A binary mask (object of class "owin" and type "mask").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

harmonise.im, compatible.im, as.im

Examples

if(require(spatstat.random)) {
   A <- setcov(square(1), dimyx=32)
   G <- as.im(function(x,y) { x^2 - y }, W=owin(), dimyx=8)
   H <- commonGrid(A, letterR, G)
   newR <- as.mask(letterR, xy=H)
   newG <- as.im(G, W=H)
   if(interactive()) plot(solist(G=newG, R=newR), main="")
}

[Package spatstat.geom version 3.3-5 Index]