compatible {spatstat.geom}	R Documentation

Test Whether Objects Are Compatible

Description

Tests whether two or more objects of the same class are compatible.

Usage

  compatible(A, B, ...)

Arguments

A,B,...
Two or more objects of the same class

Details

This generic function is used to check whether the objects A and B (and any additional objects ...) are compatible.

What is meant by ‘compatible’ depends on the class of object.

There are methods for the classes "fv", "fasp", "im" and "unitname". See the documentation for these methods for further information.

Value

Logical value: TRUE if the objects are compatible, and FALSE if they are not.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

compatible.fv, compatible.fasp, compatible.im, compatible.unitname

[Package spatstat.geom version 3.3-5 Index]