compatible.im {spatstat.geom}	R Documentation

Test Whether Pixel Images Are Compatible

Description

Tests whether two or more pixel image objects have compatible dimensions.

Usage

 ## S3 method for class 'im'
compatible(A, B, ..., tol=1e-6)

Arguments

A,B,...
Two or more pixel images (objects of class "im").

tol
Tolerance factor

Details

This function tests whether the pixel images A and B (and any additional images ...) have compatible pixel dimensions. They are compatible if they have the same number of rows and columns, the same physical pixel dimensions, and occupy the same rectangle in the plane.

The argument tol specifies the maximum tolerated error in the pixel coordinates, expressed as a fraction of the dimensions of a single pixel.

Value

Logical value: TRUE if the images are compatible, and FALSE if they are not.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

eval.im, harmonise.im, commonGrid

[Package spatstat.geom version 3.3-5 Index]