concatxy {spatstat.geom}	R Documentation

Concatenate x,y Coordinate Vectors

Description

Concatenate any number of pairs of x and y coordinate vectors.

Usage

  concatxy(...)

Arguments

...
Any number of arguments, each of which is a structure containing elements x and y.

Details

This function can be used to superimpose two or more point patterns of unmarked points (but see also superimpose which is recommended).

It assumes that each of the arguments in ... is a structure containing (at least) the elements x and y. It concatenates all the x elements into a vector x, and similarly for y, and returns these concatenated vectors.

Value

A list with two components x and y, which are the concatenations of all the corresponding x and y vectors in the argument list.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

superimpose, quadscheme

Examples

  dat <- runifrect(30)
  xy <- list(x=runif(10),y=runif(10))
  new <- concatxy(dat, xy)

[Package spatstat.geom version 3.3-5 Index]