connected.tess {spatstat.geom}	R Documentation

Connected Components of Tiles of a Tessellation

Description

Given a tessellation, find the topologically-connected pieces of each tile, and make a new tessellation using these pieces.

Usage

## S3 method for class 'tess'
connected(X, ...)

Arguments

X
A tessellation (object of class "tess").

...
Arguments passed to as.mask to determine the pixel resolution.

Details

The function connected is generic. This function connected.tess is the method for tessellations.

Given the tessellation X, the algorithm considers each tile of the tessellation, and identifies its connected components (topologically-connected pieces) using connected.owin. Each of these pieces is treated as a distinct tile and a new tessellation is made from these pieces.

The result is another tessellation obtained by subdividing each tile of X into one or more new tiles.

Value

Another tessellation (object of class "tess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

connected.owin

Examples

  BB <- grow.rectangle(Frame(letterR), 0.2)
  H <- tess(tiles=list(IN=letterR, OUT=complement.owin(letterR, BB)))
  opa <- par(mfrow=c(1,2))
  plot(H, do.col=TRUE)
  plot(connected(H), do.col=TRUE, col=2:4)
  par(opa)

[Package spatstat.geom version 3.3-5 Index]