convexhull {spatstat.geom}	R Documentation

Convex Hull

Description

Computes the convex hull of a spatial object.

Usage

convexhull(x)

Arguments

x
a window (object of class "owin"), a point pattern (object of class "ppp"), a line segment pattern (object of class "psp"), or an object that can be converted to a window by as.owin.

Details

This function computes the convex hull of the spatial object x.

Value

A window (an object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

owin, convexhull.xy, is.convex

Examples

   W <- Window(demopat)
   plot(convexhull(W), col="lightblue", border=NA)
   plot(W, add=TRUE, lwd=2)

[Package spatstat.geom version 3.3-5 Index]