convexify {spatstat.geom}	R Documentation

Weil's Convexifying Operation

Description

Converts the window W into a convex set by rearranging the edges, preserving spatial orientation of each edge.

Usage

 convexify(W, eps)

Arguments

W
A window (object of class "owin").

eps
Optional. Minimum edge length of polygonal approximation, if W is not a polygon.

Details

Weil (1995) defined a convexification operation for windows W that belong to the convex ring (that is, for any W which is a finite union of convex sets). Note that this is not the same as the convex hull.

The convexified set f(W) has the same total boundary length as W and the same distribution of orientations of the boundary. If W is a polygonal set, then the convexification f(W) is obtained by rearranging all the edges of W in order of their spatial orientation.

The argument W must be a window. If it is not already a polygonal window, it is first converted to one, using simplify.owin. The edges are sorted in increasing order of angular orientation and reassembled into a convex polygon.

Value

A window (object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

References

Weil, W. (1995) The estimation of mean particle shape and mean particle number in overlapping particle systems in the plane. Advances in Applied Probability 27, 102–119.

See Also

convexhull for the convex hull of a window.

Examples

  opa <- par(mfrow=c(1,2))
  plot(letterR)
  plot(convexify(letterR))
  par(opa)

[Package spatstat.geom version 3.3-5 Index]