corners {spatstat.geom}	R Documentation

Corners of a rectangle

Description

Returns the four corners of a rectangle

Usage

 corners(window)

Arguments

window
A window. An object of class owin, or data in any format acceptable to as.owin().

Details

This trivial function is occasionally convenient. If window is of type "rectangle" this returns the four corners of the window itself; otherwise, it returns the corners of the bounding rectangle of the window.

Value

A list with two components x and y, which are numeric vectors of length 4 giving the coordinates of the four corner points of the (bounding rectangle of the) window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

quad.object, quadscheme

Examples

  w <- unit.square()
  corners(w)
       # returns list(x=c(0,1,0,1),y=c(0,0,1,1))

[Package spatstat.geom version 3.3-5 Index]