crossdist.pp3 {spatstat.geom}	R Documentation

Pairwise distances between two different three-dimensional point patterns

Description

Computes the distances between pairs of points taken from two different three-dimensional point patterns.

Usage

  ## S3 method for class 'pp3'
crossdist(X, Y, ..., periodic=FALSE, squared=FALSE)

Arguments

X,Y
Point patterns in three dimensions (objects of class "pp3").

...
Ignored.

periodic
Logical. Specifies whether to apply a periodic edge correction.

squared
Logical. If squared=TRUE, the squared distances are returned instead (this computation is faster).

Details

Given two point patterns in three-dimensional space, this function computes the Euclidean distance from each point in the first pattern to each point in the second pattern, and returns a matrix containing these distances.

This is a method for the generic function crossdist for three-dimensional point patterns (objects of class "pp3").

This function expects two point patterns X and Y, and returns the matrix whose [i,j] entry is the distance from X[i] to Y[j].

Alternatively if periodic=TRUE, then provided the windows containing X and Y are identical and are rectangular, then the distances will be computed in the ‘periodic’ sense (also known as ‘torus’ distance): opposite edges of the rectangle are regarded as equivalent. This is meaningless if the window is not a rectangle.

Value

A matrix whose [i,j] entry is the distance from the i-th point in X to the j-th point in Y.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au based on code for two dimensions by Pavel Grabarnik pavel.grabar@issp.serpukhov.su.

See Also

crossdist, pairdist, nndist, G3est

Examples

if(require(spatstat.random)) {
   X <- runifpoint3(20)
   Y <- runifpoint3(30)
} else {
   X <- osteo$pts[[1]]
   Y <- osteo$pts[[2]]
   Y <- Y[domain(X)]
}
   d <- crossdist(X, Y)
   d <- crossdist(X, Y, periodic=TRUE)

[Package spatstat.geom version 3.3-5 Index]