crossdist.ppp {spatstat.geom}	R Documentation

Pairwise distances between two different point patterns

Description

Computes the distances between pairs of points taken from two different point patterns.

Usage

  ## S3 method for class 'ppp'
crossdist(X, Y, ...,
                periodic=FALSE, method="C", squared=FALSE,
                metric=NULL)

Arguments

X,Y
Point patterns (objects of class "ppp").

...
Ignored.

periodic
Logical. Specifies whether to apply a periodic edge correction.

method
String specifying which method of calculation to use. Values are "C" and "interpreted".

squared
Logical. If squared=TRUE, the squared distances are returned instead (this computation is faster).

metric
Optional. A distance metric (object of class "metric", see metric.object) which will be used to compute the distances.

Details

Given two point patterns, this function computes the Euclidean distance from each point in the first pattern to each point in the second pattern, and returns a matrix containing these distances.

This is a method for the generic function crossdist for point patterns (objects of class "ppp").

This function expects two point patterns X and Y, and returns the matrix whose [i,j] entry is the distance from X[i] to Y[j].

Alternatively if periodic=TRUE, then provided the windows containing X and Y are identical and are rectangular, then the distances will be computed in the ‘periodic’ sense (also known as ‘torus’ distance): opposite edges of the rectangle are regarded as equivalent. This is meaningless if the window is not a rectangle.

The argument method is not normally used. It is retained only for checking the validity of the software. If method = "interpreted" then the distances are computed using interpreted R code only. If method="C" (the default) then C code is used. The C code is faster by a factor of 4.

Value

A matrix whose [i,j] entry is the distance from the i-th point in X to the j-th point in Y.

Author(s)

Pavel Grabarnik pavel.grabar@issp.serpukhov.su and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

crossdist, crossdist.default, crossdist.psp, pairdist, nndist, Gest

Examples

   Y <- runifrect(6, Window(cells))
   d <- crossdist(cells, Y)
   d <- crossdist(cells, Y, periodic=TRUE)

[Package spatstat.geom version 3.3-5 Index]