crossdist.ppx {spatstat.geom}	R Documentation

Pairwise Distances Between Two Different Multi-Dimensional Point Patterns

Description

Computes the distances between pairs of points taken from two different multi-dimensional point patterns.

Usage

  ## S3 method for class 'ppx'
crossdist(X, Y, ...)

Arguments

X,Y
Multi-dimensional point patterns (objects of class "ppx").

...
Arguments passed to coords.ppx to determine which coordinates should be used.

Details

Given two point patterns in multi-dimensional space, this function computes the Euclidean distance from each point in the first pattern to each point in the second pattern, and returns a matrix containing these distances.

This is a method for the generic function crossdist for three-dimensional point patterns (objects of class "ppx").

This function expects two multidimensional point patterns X and Y, and returns the matrix whose [i,j] entry is the distance from X[i] to Y[j].

By default, both spatial and temporal coordinates are extracted. To obtain the spatial distance between points in a space-time point pattern, set temporal=FALSE.

Value

A matrix whose [i,j] entry is the distance from the i-th point in X to the j-th point in Y.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

crossdist, pairdist, nndist

Examples

   df <- data.frame(x=runif(3),y=runif(3),z=runif(3),w=runif(3))
   X <- ppx(data=df)
   df <- data.frame(x=runif(5),y=runif(5),z=runif(5),w=runif(5))
   Y <- ppx(data=df)
   d <- crossdist(X, Y)

[Package spatstat.geom version 3.3-5 Index]