default.image.colours {spatstat.geom}	R Documentation

Default Colours for Images in Spatstat

Description

Extract or change the default colours for images in spatstat.

Usage

default.image.colours()
reset.default.image.colours(col = NULL)

Arguments

col
A vector of colour values.

Details

These functions extract and change the current default colours used for plotting colour images in the spatstat family of packages, in particular by the functions plot.im and plot.linim.

The default colour values are a vector of character strings which can be interpreted as colours. In any particular instance of plot.im or plot.linim, the default colours are interpolated to obtain a vector of colour values of the required length (usually 256, controlled by the argument ncolours to the plot command).

default.image.colours() returns the current default colours. reset.default.image.colours(col) sets the default colours to be the vector col. reset.default.image.colours() or reset.default.image.colours(NULL) resets the factory default, which is row 29 of the Kovesi uniform perceptual contrast table described in Kovesi.

Value

A character vector of values which can be interpreted as colours.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

plot.im

Examples

 a <- default.image.colours()
 length(a)

[Package spatstat.geom version 3.3-5 Index]