default.symbolmap {spatstat.geom}	R Documentation

Default Symbol Map for Plotting a Spatial Pattern

Description

Determines the symbol map for plotting a spatial pattern, when one is not supplied by the user.

Usage

default.symbolmap(x, ...)

Arguments

x
A spatial object in the spatstat package, such as a point pattern (class "ppp").

...
Additional arguments passed to methods.

Details

In the spatstat package, an object of class "symbolmap" defines a mapping between data and graphical symbols.

If a plot command plot(x, ...) has been issued, and if the arguments were not sufficient to determine the symbol map that should be used, then default.symbolmap(x, ...) will be executed to determine the default symbol map.

The function default.symbolmap is generic, with a method for point patterns (class "ppp") and possibly for other classes.

Value

A symbol map (object of class "symbolmap").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

default.symbolmap.ppp

[Package spatstat.geom version 3.3-5 Index]