delaunay {spatstat.geom}	R Documentation

Delaunay Triangulation of Point Pattern

Description

Computes the Delaunay triangulation of a spatial point pattern.

Usage

delaunay(X)

Arguments

X
Spatial point pattern (object of class "ppp").

Details

The Delaunay triangulation of a spatial point pattern X is defined as follows. First the Dirichlet/Voronoi tessellation based on X is computed; see dirichlet. This tessellation is extended to cover the entire two-dimensional plane. Then two points of X are defined to be Delaunay neighbours if their Dirichlet/Voronoi tiles share a common boundary. Every pair of Delaunay neighbours is joined by a straight line to make the Delaunay triangulation. The result is a tessellation, consisting of disjoint triangles. The union of these triangles is the convex hull of X.

Value

A tessellation (object of class "tess"). The window of the tessellation is the convex hull of X, not the original window of X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

tess, dirichlet, convexhull.xy, ppp, delaunayDistance, delaunayNetwork.

Examples

  X <- runifrect(42)
  plot(delaunay(X))
  plot(X, add=TRUE)

[Package spatstat.geom version 3.3-5 Index]