delaunayDistance {spatstat.geom}	R Documentation

Distance on Delaunay Triangulation

Description

Computes the graph distance in the Delaunay triangulation of a point pattern.

Usage

delaunayDistance(X)

Arguments

X
Spatial point pattern (object of class "ppp").

Details

The Delaunay triangulation of a spatial point pattern X is defined as follows. First the Dirichlet/Voronoi tessellation based on X is computed; see dirichlet. This tessellation is extended to cover the entire two-dimensional plane. Then two points of X are defined to be Delaunay neighbours if their Dirichlet/Voronoi tiles share a common boundary. Every pair of Delaunay neighbours is joined by a straight line to make the Delaunay triangulation.

The graph distance in the Delaunay triangulation between two points X[i] and X[j] is the minimum number of edges of the Delaunay triangulation that must be traversed to go from X[i] to X[j]. Two points have graph distance 1 if they are immediate neighbours.

This command returns a matrix D such that D[i,j] is the graph distance between X[i] and X[j].

Value

A symmetric square matrix with non-negative integer entries.

Definition of neighbours

Note that dirichlet(X) restricts the Dirichlet tessellation to the window containing X, whereas dirichletDistance uses the Dirichlet tessellation over the entire two-dimensional plane. Some points may be Delaunay neighbours according to delaunayDistance(X) although the corresponding tiles of dirichlet(X) do not share a boundary inside Window(X).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

delaunay, delaunayNetwork.

Examples

  X <- runifrect(20)
  M <- delaunayDistance(X)
  plot(delaunay(X), lty=3)
  text(X, labels=M[1, ], cex=2)

[Package spatstat.geom version 3.3-5 Index]