diameter {spatstat.geom}	R Documentation

Diameter of an Object

Description

Computes the diameter of an object such as a two-dimensional window or three-dimensional box.

Usage

 diameter(x)

Arguments

x
A window or other object whose diameter will be computed.

Details

This function computes the diameter of an object such as a two-dimensional window or a three-dimensional box. The diameter is the maximum distance between any two points in the object.

The function diameter is generic, with methods for the class "owin" (two-dimensional windows), "box3" (three-dimensional boxes), "boxx" (multi-dimensional boxes) and "linnet" (linear networks).

Value

The numerical value of the diameter of the object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

diameter.owin, diameter.box3, diameter.boxx,

[Package spatstat.geom version 3.3-5 Index]