diameter.boxx {spatstat.geom}	R Documentation

Geometrical Calculations for Multi-Dimensional Box

Description

Calculates the volume, diameter, shortest side, side lengths, or eroded volume of a multi-dimensional box.

Usage

## S3 method for class 'boxx'
diameter(x)

## S3 method for class 'boxx'
volume(x)

## S3 method for class 'boxx'
shortside(x)

## S3 method for class 'boxx'
sidelengths(x)

## S3 method for class 'boxx'
eroded.volumes(x, r)

Arguments

x
Multi-dimensional box (object of class "boxx").

r
Numeric value or vector of numeric values for which eroded volumes should be calculated.

Details

diameter.boxx, volume.boxx and shortside.boxx compute the diameter, volume and shortest side length of the box. sidelengths.boxx returns the lengths of each side of the box.

eroded.volumes.boxx computes, for each entry r[i], the volume of the smaller box obtained by removing a slab of thickness r[i] from each face of the box. This smaller box is the subset consisting of points that lie at least r[i] units away from the boundary of the box.

Value

For diameter.boxx, shortside.boxx and volume.boxx, a single numeric value. For sidelengths.boxx, a numeric vector of length equal to the number of spatial dimensions. For eroded.volumes.boxx, a numeric vector of the same length as r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

boxx

Examples

    X <- boxx(c(0,10),c(0,10),c(0,5),c(0,2))
    diameter(X) 
    volume(X)
    shortside(X)
    sidelengths(X)
    hd <- shortside(X)/2
    eroded.volumes(X, seq(0,hd, length=10))

[Package spatstat.geom version 3.3-5 Index]