diameter.owin {spatstat.geom}	R Documentation

Diameter of a Window

Description

Computes the diameter of a window.

Usage

 ## S3 method for class 'owin'
diameter(x)

Arguments

x
A window whose diameter will be computed.

Details

This function computes the diameter of a window of arbitrary shape, i.e. the maximum distance between any two points in the window.

The argument x should be a window (an object of class "owin", see owin.object for details) or can be given in any format acceptable to as.owin().

The function diameter is generic. This function is the method for the class "owin".

Value

The numerical value of the diameter of the window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

area.owin, perimeter, edges, owin, as.owin

Examples

  w <- owin(c(0,1),c(0,1))
  diameter(w) 
  # returns sqrt(2)
  diameter(letterR)

[Package spatstat.geom version 3.3-5 Index]