dirichlet {spatstat.geom}	R Documentation

Dirichlet Tessellation of Point Pattern

Description

Computes the Dirichlet tessellation of a spatial point pattern. Also known as the Voronoi or Thiessen tessellation.

Usage

dirichlet(X)

Arguments

X
Spatial point pattern (object of class "ppp").

Details

In a spatial point pattern X, the Dirichlet tile associated with a particular point X[i] is the region of space that is closer to X[i] than to any other point in X. The Dirichlet tiles divide the two-dimensional plane into disjoint regions, forming a tessellation.

The Dirichlet tessellation is also known as the Voronoi or Thiessen tessellation.

This function computes the Dirichlet tessellation (within the original window of X) using the function deldir in the package deldir.

To ensure that there is a one-to-one correspondence between the points of X and the tiles of dirichlet(X), duplicated points in X should first be removed by X <- unique(X, rule="deldir").

The tiles of the tessellation will be computed as polygons if the original window is a rectangle or a polygon. Otherwise the tiles will be computed as binary masks.

Value

A tessellation (object of class "tess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

tess, delaunay, ppp, dirichletVertices.

For the Dirichlet tessellation on a linear network, see lineardirichlet.

Examples

  X <- runifrect(42)
  plot(dirichlet(X))
  plot(X, add=TRUE)

[Package spatstat.geom version 3.3-5 Index]