dirichletVertices {spatstat.geom}	R Documentation

Vertices and Edges of Dirichlet Tessellation

Description

Computes the Dirichlet-Voronoi tessellation of a point pattern and extracts the vertices or edges of the tiles.

Usage

dirichletVertices(X)

dirichletEdges(X, clip=TRUE)

Arguments

X
Point pattern (object of class "ppp").

clip
Logical value specifying whether to clip the tile edges to the window. See Details.

Details

These function compute the Dirichlet-Voronoi tessellation of X (see dirichlet) and extract the vertices or edges of the tiles of the tessellation.

The Dirichlet vertices are the spatial locations which are locally farthest away from X, that is, where the distance function of X reaches a local maximum.

The Dirichlet edges are the dividing lines equally distant between a pair of points of X.

The Dirichlet tessellation of X is computed using dirichlet. The vertices or edges of all tiles of the tessellation are extracted.

For dirichletVertices, any vertex which lies on the boundary of the window of X is deleted. The remaining vertices are returned, as a point pattern, without duplicated entries.

For dirichletEdges, the edges are initially computed inside the rectangle Frame(X). Then if clip=TRUE (the default), these edges are intersected with Window(X), which may cause an edge to be broken into several pieces.

Value

dirichletVertices returns a point pattern (object of class "ppp") in the same window as X.

dirichletEdges returns a line segment pattern (object of class "psp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

dirichlet, dirichletAreas

Examples

  plot(dirichlet(cells))

  plot(dirichletVertices(cells), add=TRUE)

  ed <- dirichletEdges(cells)

[Package spatstat.geom version 3.3-5 Index]