dirichletWeights {spatstat.geom}	R Documentation

Compute Quadrature Weights Based on Dirichlet Tessellation

Description

Computes quadrature weights for a given set of points, using the areas of tiles in the Dirichlet tessellation.

Usage

 dirichletWeights(X, window=NULL, exact=TRUE, ...)

Arguments

X
Data defining a point pattern.

window
Default window for the point pattern

exact
Logical value. If TRUE, compute exact areas using the package deldir. If FALSE, compute approximate areas using a pixel raster.

...
Ignored.

Details

This function computes a set of quadrature weights for a given pattern of points (typically comprising both “data” and 'dummy” points). See quad.object for an explanation of quadrature weights and quadrature schemes.

The weights are computed using the Dirichlet tessellation. First X and (optionally) window are converted into a point pattern object. Then the Dirichlet tessellation of the points of X is computed. The weight attached to a point of X is the area of its Dirichlet tile (inside the window Window(X)).

If exact=TRUE the Dirichlet tessellation is computed exactly by the Lee-Schachter algorithm using the package deldir. Otherwise a pixel raster approximation is constructed and the areas are approximations to the true weights. In all cases the sum of the weights is equal to the area of the window.

Value

Vector of nonnegative weights for each point in X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

quad.object, gridweights

Examples

  Q <- quadscheme(runifrect(10))
  X <- as.ppp(Q) # data and dummy points together
  w <- dirichletWeights(X, exact=FALSE)

[Package spatstat.geom version 3.3-5 Index]