distmap {spatstat.geom}	R Documentation

Distance Map

Description

Compute the distance map of an object, and return it as a pixel image. Generic.

Usage

  distmap(X, ...)

Arguments

X
Any suitable dataset representing a two-dimensional object, such as a point pattern (object of class "ppp"), a window (object of class "owin") or a line segment pattern (object of class "psp").

...
Arguments passed to as.mask to control pixel resolution.

Details

The “distance map” of a set of points A is the function f whose value f(x) is defined for any two-dimensional location x as the shortest distance from x to A.

This function computes the distance map of the set X and returns the distance map as a pixel image.

This is generic. Methods are provided for point patterns (distmap.ppp), line segment patterns (distmap.psp) and windows (distmap.owin) as well as other classes.

Value

A pixel image (object of class "im") whose grey scale values are the values of the distance map.

Distance values

The pixel values in the image distmap(X) are distances, expressed as multiples of the unit of length of the spatial coordinates in X. The unit of length is given by unitname(X).

Note that, if the unit of length in X is a composite expression such as ‘2 microns’, then the values in distmap(X) are expressed as multiples of 2 microns, rather than being expressed in microns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

distmap.ppp, distmap.psp, distmap.owin, distfun

Examples

  U <- distmap(cells)
  V <- distmap(letterR)
  if(interactive()) {
  plot(U)
  plot(V)
  }

[Package spatstat.geom version 3.3-5 Index]