domain {spatstat.geom}	R Documentation

Extract the Domain of any Spatial Object

Description

Given a spatial object such as a point pattern, in any number of dimensions, this function extracts the spatial domain in which the object is defined.

Usage

   domain(X, ...)

   ## S3 method for class 'ppp'
domain(X, ...)

   ## S3 method for class 'psp'
domain(X, ...)

   ## S3 method for class 'im'
domain(X, ...)

   ## S3 method for class 'ppx'
domain(X, ...)

   ## S3 method for class 'pp3'
domain(X, ...)

 ## S3 method for class 'quad'
domain(X, ...)

 ## S3 method for class 'quadratcount'
domain(X, ...)

 ## S3 method for class 'tess'
domain(X, ...)

 ## S3 method for class 'layered'
domain(X, ...)

 ## S3 method for class 'distfun'
domain(X, ...)

 ## S3 method for class 'nnfun'
domain(X, ...)

 ## S3 method for class 'funxy'
domain(X, ...)

Arguments

X
A spatial object such as a point pattern (in any number of dimensions), line segment pattern or pixel image.

...
Extra arguments. They are ignored by all the methods listed here.

Details

The function domain is generic.

For a spatial object X in any number of dimensions, domain(X) extracts the spatial domain in which X is defined.

For a two-dimensional object X, typically domain(X) is the same as Window(X).

Exceptions occur for methods related to linear networks.

Value

A spatial object representing the domain of X. Typically a window (object of class "owin"), a three-dimensional box ("box3"), a multidimensional box ("boxx") or a linear network ("linnet").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

domain.ppm, domain.quadrattest, domain.rmhmodel, domain.lpp. Window, Frame.

Examples

  domain(redwood)
  domain(bei.extra$elev)
  domain(osteo$pts[[1]])

[Package spatstat.geom version 3.3-5 Index]