edit.hyperframe {spatstat.geom}	R Documentation

Invoke Text Editor on Hyperframe

Description

Invokes a text editor allowing the user to inspect and change entries in a hyperframe.

Usage

## S3 method for class 'hyperframe'
edit(name, ...)

Arguments

name
A hyperframe (object of class "hyperframe").

...
Other arguments passed to edit.data.frame.

Details

The function edit is generic. This function is the methods for objects of class "hyperframe".

The hyperframe name is converted to a data frame or array, and the text editor is invoked. The user can change entries in the columns of data, and create new columns of data.

Only the columns of atomic data (numbers, characters, factor values etc) can be edited.

Note that the original object name is not changed; the function returns the edited dataset.

Value

Another hyperframe.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

edit.data.frame, edit.ppp

Examples

  if(interactive()) Z <- edit(flu)

[Package spatstat.geom version 3.3-5 Index]