edit.ppp {spatstat.geom}	R Documentation

Invoke Text Editor on Spatial Data

Description

Invokes a text editor allowing the user to inspect and change entries in a spatial dataset.

Usage

## S3 method for class 'ppp'
edit(name, ...)

## S3 method for class 'psp'
edit(name, ...)

## S3 method for class 'im'
edit(name, ...)

Arguments

name
A spatial dataset (object of class "ppp", "psp" or "im").

...
Other arguments passed to edit.data.frame.

Details

The function edit is generic. These functions are methods for spatial objects of class "ppp", "psp" and "im".

The spatial dataset name is converted to a data frame or array, and the text editor is invoked. The user can change the values of spatial coordinates or marks of the points in a point pattern, or the coordinates or marks of the segments in a segment pattern, or the pixel values in an image. The names of the columns of marks can also be edited.

If name is a pixel image, it is converted to a matrix and displayed in the same spatial orientation as if the image had been plotted.

Note that the original object name is not changed; the function returns the edited dataset.

Value

Object of the same kind as name containing the edited data.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

edit.data.frame, edit.hyperframe

Examples

  if(interactive()) Z <- edit(cells)

[Package spatstat.geom version 3.3-5 Index]