eroded.areas {spatstat.geom}	R Documentation

Areas of Morphological Erosions

Description

Computes the areas of successive morphological erosions of a window.

Usage

 eroded.areas(w, r, subset=NULL)

Arguments

w
A window.

r
Numeric vector of radii at which erosions will be performed.

subset
Optional window inside which the areas should be computed.

Details

This function computes the areas of the erosions of the window w by each of the radii r[i].

The morphological erosion of a set W by a distance r > 0 is the subset consisting of all points x in W such that the distance from x to the boundary of W is greater than or equal to r. In other words it is the result of trimming a margin of width r off the set W.

The argument r should be a vector of positive numbers. The argument w should be a window (an object of class "owin", see owin.object for details) or can be given in any format acceptable to as.owin().

Unless w is a rectangle, the computation is performed using a pixel raster approximation.

To compute the eroded window itself, use erosion.

Value

Numeric vector, of the same length as r, giving the areas of the successive erosions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

owin, as.owin, erosion

Examples

  w <- owin(c(0,1),c(0,1))
  a <- eroded.areas(w, seq(0.01,0.49,by=0.01))

[Package spatstat.geom version 3.3-5 Index]