extrapolate.psp {spatstat.geom}	R Documentation

Extrapolate Line Segments to Obtain Infinite Lines

Description

Given a spatial pattern of line segments, extrapolate the segments to infinite lines.

Usage

extrapolate.psp(x, ...)

Arguments

x
Spatial pattern of line segments (object of class "psp").

...
Ignored.

Details

Each line segment in the pattern x is extrapolated to an infinite line, drawn through its two endpoints. The resulting pattern of infinite lines is returned as an object of class "infline".

If a segment's endpoints are identical (so that it has zero length) the resulting infinite line is vertical (i.e. parallel to the y coordinate axis).

Value

An object of class "infline" representing the pattern of infinite lines. See infline for details of structure.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

psp, infline

midpoints.psp, lengths_psp angles.psp, endpoints.psp.

Examples

   X <- psp(runif(4), runif(4), runif(4), runif(4), window=owin())
   Y <- extrapolate.psp(X)
   plot(X, col=3, lwd=4)
   plot(Y, lty=3)
   Y

[Package spatstat.geom version 3.3-5 Index]