fardist {spatstat.geom}	R Documentation

Farthest Distance to Boundary of Window

Description

Computes the farthest distance from each pixel, or each data point, to the boundary of the window.

Usage

  fardist(X, ...)

  ## S3 method for class 'owin'
fardist(X, ..., squared=FALSE)

  ## S3 method for class 'ppp'
fardist(X, ..., squared=FALSE)

Arguments

X
A spatial object such as a window or point pattern.

...
Arguments passed to as.mask to determine the pixel resolution, if required.

squared
Logical. If TRUE, the squared distances will be returned.

Details

The function fardist is generic, with methods for the classes owin and ppp.

For a window W, the command fardist(W) returns a pixel image in which the value at each pixel is the largest distance from that pixel to the boundary of W.

For a point pattern X, with window W, the command fardist(X) returns a numeric vector with one entry for each point of X, giving the largest distance from that data point to the boundary of W.

Value

For fardist.owin, a pixel image (object of class "im").

For fardist.ppp, a numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

Examples

  fardist(cells)

  plot(FR <- fardist(letterR))

[Package spatstat.geom version 3.3-5 Index]