gridweights {spatstat.geom}	R Documentation

Compute Quadrature Weights Based on Grid Counts

Description

Computes quadrature weights for a given set of points, using the “counting weights” for a grid of rectangular tiles.

Usage

 gridweights(X, ntile, ..., window=NULL, verbose=FALSE, npix=NULL, areas=NULL)

Arguments

X
Data defining a point pattern.

ntile
Number of tiles in each row and column of the rectangular grid. An integer vector of length 1 or 2.

...
Ignored.

window
Default window for the point pattern

verbose
Logical flag. If TRUE, information will be printed about the computation of the grid weights.

npix
Dimensions of pixel grid to use when computing a digital approximation to the tile areas.

areas
Vector of areas of the tiles, if they are already known.

Details

This function computes a set of quadrature weights for a given pattern of points (typically comprising both “data” and 'dummy” points). See quad.object for an explanation of quadrature weights and quadrature schemes.

The weights are computed by the “counting weights” rule based on a regular grid of rectangular tiles. First X and (optionally) window are converted into a point pattern object. Then the bounding rectangle of the window of the point pattern is divided into a regular ntile[1] * ntile[2] grid of rectangular tiles. The weight attached to a point of X is the area of the tile in which it lies, divided by the number of points of X lying in that tile.

For non-rectangular windows the tile areas are currently calculated by approximating the window as a binary mask. The accuracy of this approximation is controlled by npix, which becomes the argument dimyx of as.mask.

Value

Vector of nonnegative weights for each point in X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

quad.object, dirichletWeights

Examples

  Q <- quadscheme(runifrect(15))
  X <- as.ppp(Q) # data and dummy points together
  w <- gridweights(X, 10)
  w <- gridweights(X, c(10, 10))

[Package spatstat.geom version 3.3-5 Index]