grow.boxx {spatstat.geom}	R Documentation

Add margins to box in any dimension

Description

Adds a margin to a box of class boxx.

Usage

 grow.boxx(W, left, right = left)
 grow.box3(W, left, right = left)

Arguments

W
A box (object of class "boxx" or "box3").

left
Width of margin to be added to left endpoint of box side in every dimension. A single nonnegative number, or a vector of same length as the dimension of the box to add different left margin in each dimension.

right
Width of margin to be added to right endpoint of box side in every dimension. A single nonnegative number, or a vector of same length as the dimension of the box to add different right margin in each dimension.

Value

Another object of the same class "boxx" or "box3" representing the window after margins are added.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

grow.rectangle, boxx, box3

Examples

  w <- boxx(c(0,10), c(0,10), c(0,10), c(0,10))
  # add a margin of size 1 on both sides in all four dimensions
  b12 <- grow.boxx(w, 1)

  # add margin of size 2 at left, and margin of size 3 at right,
  # in each dimension.
  v <- grow.boxx(w, 2, 3)

[Package spatstat.geom version 3.3-5 Index]