harmonise {spatstat.geom}	R Documentation

Make Objects Compatible

Description

Converts several objects of the same class to a common format so that they can be combined or compared.

Usage

harmonise(...)
harmonize(...)

Arguments

...
Any number of objects of the same class.

Details

This generic command takes any number of objects of the same class, and attempts to make them compatible in the sense of compatible so that they can be combined or compared.

There are methods for the classes "fv" (harmonise.fv) and "im" (harmonise.im).

All arguments ... must be objects of the same class. The result will be a list, of length equal to the number of arguments ..., containing new versions of each of these objects, converted to a common format. If the arguments were named (name=value) then the return value also carries these names.

Value

A list, of length equal to the number of arguments ..., whose entries are objects of the same class. If the arguments were named (name=value) then the return value also carries these names.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

compatible, harmonise.fv, harmonise.im

[Package spatstat.geom version 3.3-5 Index]