harmonise.owin {spatstat.geom}	R Documentation

Make Windows Compatible

Description

Convert several windows to a common pixel raster.

Usage

## S3 method for class 'owin'
harmonise(...)

## S3 method for class 'owin'
harmonize(...)

Arguments

...
Any number of windows (objects of class "owin") or data which can be converted to windows by as.owin.

Details

This function makes any number of windows compatible, by converting them all to a common pixel grid.

This only has an effect if one of the windows is a binary mask. If all the windows are rectangular or polygonal, they are returned unchanged.

The command harmonise is generic. This is the method for objects of class "owin".

Each argument must be a window (object of class "owin"), or data that can be converted to a window by as.owin.

The common pixel grid is determined by inspecting all the windows in the argument list, computing the bounding box of all the windows, then finding the binary mask with the finest spatial resolution, and extending its pixel grid to cover the bounding box.

The return value is a list with entries corresponding to the input arguments. If the arguments were named (name=value) then the return value also carries these names.

If you just want to determine the appropriate pixel resolution, without converting the windows, use commonGrid.

Value

A list of windows, of length equal to the number of arguments .... The list belongs to the class "solist".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

commonGrid, harmonise.im, as.owin

Examples

   harmonise(X=letterR,
             Y=grow.rectangle(Frame(letterR), 0.2),
             Z=as.mask(letterR, eps=0.1),
             V=as.mask(letterR, eps=0.07))

[Package spatstat.geom version 3.3-5 Index]