imcov {spatstat.geom}	R Documentation

Spatial Covariance of a Pixel Image

Description

Computes the unnormalised spatial covariance function of a pixel image.

Usage

 imcov(X, Y=X)

Arguments

X
A pixel image (object of class "im".

Y
Optional. Another pixel image.

Details

The (uncentred, unnormalised) spatial covariance function of a pixel image X in the plane is the function C(v) defined for each vector v as

C(v) = integral of X(u) * X(u-v) du

where the integral is over all spatial locations u, and where X(u) denotes the pixel value at location u.

This command computes a discretised approximation to the spatial covariance function, using the Fast Fourier Transform. The return value is another pixel image (object of class "im") whose greyscale values are values of the spatial covariance function.

If the argument Y is present, then imcov(X,Y) computes the set cross-covariance function C(u) defined as

C(v) = integral of X(u) * Y(u-v) du.

Note that imcov(X,Y) is equivalent to convolve.im(X,Y,reflectY=TRUE).

Value

A pixel image (an object of class "im") representing the spatial covariance function of X, or the cross-covariance of X and Y.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

setcov, convolve.im, owin, as.owin, erosion

Examples

  X <- as.im(square(1))
  v <- imcov(X)
  plot(v)

[Package spatstat.geom version 3.3-5 Index]