intensity.ppx {spatstat.geom}	R Documentation

Intensity of a Multidimensional Space-Time Point Pattern

Description

Calculates the intensity of points in a multi-dimensional point pattern of class "ppx" or "pp3".

Usage

  ## S3 method for class 'ppx'
intensity(X, ...)

Arguments

X
Point pattern of class "ppx" or "pp3".

...
Ignored.

Details

This is a method for the generic function intensity. It computes the empirical intensity of a multi-dimensional point pattern (object of class "ppx" including "pp3"), i.e. the average density of points per unit volume.

If the point pattern is multitype, the intensities of the different types are computed separately.

Value

A single number or a numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

Examples

  X <- osteo$pts[[1]]
  intensity(X)
  marks(X) <- factor(sample(letters[1:3], npoints(X), replace=TRUE))
  intensity(X)

[Package spatstat.geom version 3.3-5 Index]