intensity.quadratcount {spatstat.geom}	R Documentation

Intensity Estimates Using Quadrat Counts

Description

Uses quadrat count data to estimate the intensity of a point pattern in each tile of a tessellation, assuming the intensity is constant in each tile.

Usage

## S3 method for class 'quadratcount'
intensity(X, ..., image=FALSE)

Arguments

X
An object of class "quadratcount".

image
Logical value specifying whether to return a table of estimated intensities (the default) or a pixel image of the estimated intensity (image=TRUE).

...
Arguments passed to as.mask to determine the resolution of the pixel image, if image=TRUE.

Details

This is a method for the generic function intensity. It computes an estimate of the intensity of a point pattern from its quadrat counts.

The argument X should be an object of class "quadratcount". It would have been obtained by applying the function quadratcount to a point pattern (object of class "ppp"). It contains the counts of the numbers of points of the point pattern falling in each tile of a tessellation.

Using this information, intensity.quadratcount divides the quadrat counts by the tile areas, yielding the average density of points per unit area in each tile of the tessellation.

If image=FALSE (the default), these intensity values are returned in a contingency table. Cells of the contingency table correspond to tiles of the tessellation.

If image=TRUE, the estimated intensity function is returned as a pixel image. For each pixel, the pixel value is the estimated intensity in the tile which contains that pixel.

Value

If image=FALSE (the default), a contingency table. If image=TRUE, a pixel image (object of class "im").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

intensity, quadratcount

Examples

  qa <- quadratcount(swedishpines, 4,3)
  qa
  intensity(qa)
  plot(intensity(qa, image=TRUE))

[Package spatstat.geom version 3.3-5 Index]