is.boxx {spatstat.geom}	R Documentation

Recognise a Multi-Dimensional Box

Description

Checks whether its argument is a multidimensional box (object of class "boxx").

Usage

  is.boxx(x)

Arguments

x
Any object.

Details

This function tests whether the object x is a multidimensional box of class "boxx".

The result is determined to be TRUE if x inherits from "boxx", i.e. if x has "boxx" amongst its classes.

Value

A logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

methods.boxx, boxx.

Examples

   B <- boxx(c(0,10),c(0,10),c(0,5),c(0,1), unitname="km")
   is.boxx(B)
   is.boxx(42)

[Package spatstat.geom version 3.3-5 Index]