is.connected.ppp {spatstat.geom}	R Documentation

Determine Whether a Point Pattern is Connected

Description

Determine whether a point pattern is topologically connected when all pairs of points closer than a threshold distance are joined.

Usage

## S3 method for class 'ppp'
is.connected(X, R, ...)

Arguments

X
A point pattern (object of class "ppp").

R
Threshold distance. Pairs of points closer than R units apart will be joined together.

...
Ignored.

Details

The function is.connected is generic. This is the method for point patterns (objects of class "ppp").

The point pattern X is first converted into an abstract graph by joining every pair of points that lie closer than R units apart. Then the algorithm determines whether this graph is connected.

That is, the result of is.connected(X) is TRUE if any point in X can be reached from any other point, by a series of steps between points of X, each step being shorter than R units in length.

Value

A logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

is.connected, connected.ppp.

Examples

  is.connected(redwoodfull, 0.1)
  is.connected(redwoodfull, 0.2)

[Package spatstat.geom version 3.3-5 Index]