is.convex {spatstat.geom}	R Documentation

Test Whether a Window is Convex

Description

Determines whether a window is convex.

Usage

  is.convex(x)

Arguments

x
Window (object of class "owin").

Details

If x is a rectangle, the result is TRUE.

If x is polygonal, the result is TRUE if x consists of a single polygon and this polygon is equal to the minimal convex hull of its vertices computed by chull.

If x is a mask, the algorithm first extracts all boundary pixels of x using vertices. Then it computes the (polygonal) convex hull K of the boundary pixels. The result is TRUE if every boundary pixel lies within one pixel diameter of an edge of K.

Value

Logical value, equal to TRUE if x is convex.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

owin, convexhull.xy, vertices

[Package spatstat.geom version 3.3-5 Index]