is.empty {spatstat.geom}	R Documentation

Test Whether An Object Is Empty

Description

Checks whether the argument is an empty window, an empty point pattern, etc.

Usage

is.empty(x)
## S3 method for class 'owin'
is.empty(x)
## S3 method for class 'ppp'
is.empty(x)
## S3 method for class 'psp'
is.empty(x)
## Default S3 method:
is.empty(x)

Arguments

x
A window (object of class "owin"), a point pattern (object of class "ppp"), or a line segment pattern (object of class "psp").

Details

This function tests whether the object x represents an empty spatial object, such as an empty window, a point pattern with zero points, or a line segment pattern with zero line segments.

An empty window can be obtained as the output of intersect.owin, erosion, opening, complement.owin and some other operations.

An empty point pattern or line segment pattern can be obtained as the result of simulation.

Value

Logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

[Package spatstat.geom version 3.3-5 Index]