is.lpp {spatstat.geom}	R Documentation

Test Whether An Object Is A Point Pattern on a Linear Network

Description

Checks whether its argument is a point pattern on a linear network (object of class "lpp").

Usage

is.lpp(x)

Arguments

x
Any object.

Details

This function tests whether the object x is a point pattern object of class "lpp".

Value

TRUE if x is a point pattern of class "lpp", otherwise FALSE.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

[Package spatstat.geom version 3.3-5 Index]