is.ppp {spatstat.geom}	R Documentation

Test Whether An Object Is A Point Pattern

Description

Checks whether its argument is a point pattern (object of class "ppp").

Usage

is.ppp(x)

Arguments

x
Any object.

Details

This function tests whether the object x is a point pattern object of class "ppp". See ppp.object for details of this class.

The result is determined to be TRUE if x inherits from "ppp", i.e. if x has "ppp" amongst its classes.

Value

TRUE if x is a point pattern, otherwise FALSE.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

[Package spatstat.geom version 3.3-5 Index]