is.rectangle {spatstat.geom}	R Documentation

Determine Type of Window

Description

Determine whether a window is a rectangle, a polygonal region, or a binary mask.

Usage

  is.rectangle(w)
  is.polygonal(w)
  is.mask(w)

Arguments

w
Window to be inspected. An object of class "owin".

Details

These simple functions determine whether a window w (object of class "owin") is a rectangle (is.rectangle(w) = TRUE), a domain with polygonal boundary (is.polygonal(w) = TRUE), or a binary pixel mask (is.mask(w) = TRUE).

Value

Logical value, equal to TRUE if w is a window of the specified type.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

owin

[Package spatstat.geom version 3.3-5 Index]