layerplotargs {spatstat.geom}	R Documentation

Extract or Replace the Plot Arguments of a Layered Object

Description

Extracts or replaces the plot arguments of a layered object.

Usage

layerplotargs(L)

layerplotargs(L) <- value

Arguments

L
An object of class "layered" created by the function layered.

value
Replacement value. A list, with the same length as L, whose elements are lists of plot arguments.

Details

These commands extract or replace the plotargs in a layered object. See layered.

The replacement value should normally have the same length as the current value. However, it can also be a list with one element which is a list of parameters. This will be replicated to the required length.

For the assignment function layerplotargs<-, the argument L can be any spatial object; it will be converted to a layered object with a single layer.

Value

layerplotargs returns a list of lists of plot arguments.

"layerplotargs<-" returns the updated object of class "layered".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

layered, methods.layered, [.layered.

Examples

   W <- square(2)
   L <- layered(W=W, X=cells)
   ## The following are equivalent
   layerplotargs(L) <- list(list(), list(pch=16))
   layerplotargs(L)[[2]] <- list(pch=16)
   layerplotargs(L)$X <- list(pch=16)

   ## The following are equivalent
   layerplotargs(L) <- list(list(cex=2), list(cex=2))
   layerplotargs(L) <- list(list(cex=2))

[Package spatstat.geom version 3.3-5 Index]