layout.boxes {spatstat.geom}	R Documentation

Generate a Row or Column Arrangement of Rectangles.

Description

A simple utility to generate a row or column of boxes (rectangles) for use in point-and-click panels.

Usage

layout.boxes(B, n, horizontal = FALSE, aspect = 0.5, usefrac = 0.9)

Arguments

B
Bounding rectangle for the boxes. An object of class "owin".

n
Integer. The number of boxes.

horizontal
Logical. If TRUE, arrange the boxes in a horizontal row. If FALSE (the default), arrange them in a vertical column.

aspect
A single finite positive number, giving the aspect ratio (height divided by width) of each box, or NA or Inf, indicating that the aspect ratio is unconstrained.

usefrac
Number between 0 and 1. The fraction of height or width of B that should be occupied by boxes.

Details

This simple utility generates a list of boxes (rectangles) inside the bounding box B arranged in a regular row or column. It is useful for generating the positions of the panel buttons in the function simplepanel.

The argument aspect specifies the ratio of height to width (height divided by width). If aspect is a finite numerical value, then the boxes will have the given aspect ratio. If aspect is Inf or NA, aspect ratio is unconstrained; the boxes will have the maximum possible width and height.

Value

A list of rectangles (objects of class "owin" which are rectangles).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

simplepanel

Examples

  B <- owin(c(0,10),c(0,1))
  boxes <- layout.boxes(B, 5, horizontal=TRUE)
  plot(B, main="", col="blue")
  niets <- lapply(boxes, plot, add=TRUE, col="grey")

[Package spatstat.geom version 3.3-5 Index]