marks.tess {spatstat.geom}	R Documentation

Marks of a Tessellation

Description

Extract or change the marks attached to the tiles of a tessellation.

Usage

## S3 method for class 'tess'
marks(x, ...)           

## S3 replacement method for class 'tess'
marks(x, ...) <- value

## S3 method for class 'tess'
unmark(X)

Arguments

x,X
Tessellation (object of class "tess")

...
Ignored.

value
Vector or data frame of mark values, or NULL.

Details

These functions extract or change the marks attached to each of the tiles in the tessellation x. They are methods for the generic functions marks, marks<- and unmark for the class "tess" of tessellations

The expression marks(x) extracts the marks of x. The assignment marks(x) <- value assigns new marks to the dataset x, and updates the dataset x in the current environment.

The marks can be a vector, a factor, a data frame or a hyperframe.

For the assignment marks(x) <- value, the value should be a vector or factor of length equal to the number of tiles in x, or a data frame or hyperframe with as many rows as there are tiles in x. If value is a single value, or a data frame or hyperframe with one row, then it will be replicated so that the same marks will be attached to each tile.

To remove marks, use marks(x) <- NULL or unmark(x).

Value

For marks(x), the result is a vector, factor, data frame or hyperframe, containing the mark values attached to the tiles of x. If there are no marks, the result is NULL.

For unmark(x), the result is the tessellation without marks.

For marks(x) <- value, the result is the updated tessellation x (with the side-effect that the dataset x is updated in the current environment).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

marks, marks<-

Examples

  D <- dirichlet(cells)
  marks(D) <- tile.areas(D)

[Package spatstat.geom version 3.3-5 Index]