methods.box3 {spatstat.geom}	R Documentation

Methods for Three-Dimensional Box

Description

Methods for class "box3".

Usage

  ## S3 method for class 'box3'
print(x, ...)
  ## S3 method for class 'box3'
unitname(x) 
  ## S3 replacement method for class 'box3'
unitname(x) <- value

Arguments

x
Object of class "box3" representing a three-dimensional box.

...
Other arguments passed to print.default.

value
Name of the unit of length. See unitname.

Details

These are methods for the generic functions print and unitname for the class "box3" of three-dimensional boxes.

The print method prints a description of the box, while the unitname method extracts the name of the unit of length in which the box coordinates are expressed.

Value

For print.box3 the value is NULL. For unitname.box3 an object of class "units".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

box3, print, unitname

Examples

   X <- box3(c(0,10),c(0,10),c(0,5), unitname=c("metre", "metres"))
   X
   unitname(X)
   # Northern European usage
   unitname(X) <- "meter"

[Package spatstat.geom version 3.3-5 Index]