methods.boxx {spatstat.geom}	R Documentation

Methods for Multi-Dimensional Box

Description

Methods for class "boxx".

Usage

  ## S3 method for class 'boxx'
print(x, ...)
  ## S3 method for class 'boxx'
unitname(x)
  ## S3 replacement method for class 'boxx'
unitname(x) <- value
  ## S3 method for class 'boxx'
scale(x, center=TRUE, scale=TRUE)

Arguments

x
Object of class "boxx" representing a multi-dimensional box.

...
Other arguments passed to print.default.

value
Name of the unit of length. See unitname.

center,scale
Arguments passed to scale.default to determine the rescaling.

Details

These are methods for the generic functions print, unitname, unitname<- and scale for the class "boxx" of multi-dimensional boxes.

The print method prints a description of the box, the unitname method extracts the name of the unit of length in which the box coordinates are expressed, while the assignment method for unitname assigns this unit name.

The scale method rescales each spatial coordinate of x.

Value

For print.boxx the value is NULL. For unitname.boxx an object of class "units". For unitname<-.boxx and scale.boxx the result is the updated "boxx" object x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

boxx, is.boxx, print, unitname, scale

Examples

   B <- boxx(c(0,10),c(0,10),c(0,5),c(0,1), unitname=c("metre", "metres"))
   B
   is.boxx(B)
   unitname(B)
   # Northern European usage
   unitname(B) <- "meter"
   scale(B)

[Package spatstat.geom version 3.3-5 Index]