methods.distfun {spatstat.geom}	R Documentation

Geometrical Operations for Distance Functions

Description

Methods for objects of the class "distfun".

Usage

## S3 method for class 'distfun'
shift(X, ...)

## S3 method for class 'distfun'
rotate(X, ...)

## S3 method for class 'distfun'
scalardilate(X, ...)

## S3 method for class 'distfun'
affine(X, ...)

## S3 method for class 'distfun'
flipxy(X)

## S3 method for class 'distfun'
reflect(X)

## S3 method for class 'distfun'
rescale(X, s, unitname)

Arguments

X
Object of class "distfun" representing the distance function of a spatial object.

...
Arguments passed to the next method for the geometrical operation. See Details.

s, unitname
Arguments passed to the next method for rescale.

Details

These are methods for the generic functions shift, rotate, scalardilate, affine, flipxy and reflect which perform geometrical operations on spatial objects, and for the generic rescale which changes the unit of length.

The argument X should be an object of class "distfun" representing the distance function of a spatial object Y. Objects of class "distfun" are created by distfun.

The methods apply the specified geometrical transformation to the original object Y, producing a new object Z of the same type as Y. They then create a new distfun object representing the distance function of Z.

Value

Another object of class "distfun".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

distfun, methods.funxy.

Examples

   (f <- distfun(letterR))
   plot(f)
   flipxy(f)
   shift(f, origin="midpoint")
   plot(rotate(f, angle=pi/2))

   (g <- distfun(lansing))
   rescale(g)

[Package spatstat.geom version 3.3-5 Index]