methods.funxy {spatstat.geom}	R Documentation

Methods for Spatial Functions

Description

Methods for objects of the class "funxy".

Usage

## S3 method for class 'funxy'
contour(x, ...)
## S3 method for class 'funxy'
persp(x, ...)
## S3 method for class 'funxy'
plot(x, ...)

Arguments

x
Object of class "funxy" representing a function of x,y coordinates.

...
Named arguments controlling the plot. See Details.

Details

These are methods for the generic functions plot, contour and persp for the class "funxy" of spatial functions.

Objects of class "funxy" are created, for example, by the commands distfun and funxy.

The plot, contour and persp methods first convert x to a pixel image object using as.im, then display it using plot.im, contour.im or persp.im.

Additional arguments ... are either passed to as.im.function to control the spatial resolution of the pixel image, or passed to contour.im, persp.im or plot.im to control the appearance of the plot.

Value

NULL.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

funxy, distfun, as.im, plot.im, persp.im, contour.im, spatstat.options

Examples

   f <- distfun(letterR)
   contour(f)
   B <- owin(c(1,5), c(-1, 4))
   contour(f, W=B)
   persp(f, W=B, theta=40, phi=40, border=NA, shade=0.7)

[Package spatstat.geom version 3.3-5 Index]