methods.pp3 {spatstat.geom}	R Documentation

Methods for three-dimensional point patterns

Description

Methods for class "pp3".

Usage

  ## S3 method for class 'pp3'
print(x, ...)
  ## S3 method for class 'summary.pp3'
print(x, ...)
  ## S3 method for class 'pp3'
summary(object, ...)
  ## S3 method for class 'pp3'
unitname(x)
  ## S3 replacement method for class 'pp3'
unitname(x) <- value

Arguments

x,object
Object of class "pp3".

...
Ignored.

value
Name of the unit of length. See unitname.

Details

These are methods for the generic functions print, summary, unitname and unitname<- for the class "pp3" of three-dimensional point patterns.

The print and summary methods print a description of the point pattern.

The unitname method extracts the name of the unit of length in which the point coordinates are expressed. The unitname<- method assigns the name of the unit of length.

Value

For print.pp3 the value is NULL. For unitname.pp3 an object of class "units".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

pp3, print, unitname unitname<-

Examples

   X <- pp3(runif(42),runif(42),runif(42), box3(c(0,1), unitname="mm"))
   X
   unitname(X)
   unitname(X) <- c("foot", "feet")
   summary(X)

[Package spatstat.geom version 3.3-5 Index]