methods.ppx {spatstat.geom}	R Documentation

Methods for Multidimensional Space-Time Point Patterns

Description

Methods for printing and plotting a general multidimensional space-time point pattern.

Usage

## S3 method for class 'ppx'
print(x, ...)
## S3 method for class 'ppx'
plot(x, ...)
## S3 method for class 'ppx'
unitname(x)
## S3 replacement method for class 'ppx'
unitname(x) <- value
## S3 method for class 'ppx'
scale(x, center=TRUE, scale=TRUE)

Arguments

x
Multidimensional point pattern (object of class "ppx").

...
Additional arguments passed to plot methods.

value
Name of the unit of length. See unitname.

center,scale
Arguments passed to scale.default to determine the rescaling.

Details

These are methods for the generic functions print, plot, unitname, unitname<- and scale for the class "ppx" of multidimensional point patterns.

The print method prints a description of the point pattern and its spatial domain.

The unitname method extracts the name of the unit of length in which the point coordinates are expressed. The unitname<- method assigns the name of the unit of length.

The scale method rescales each spatial coordinate of x.

Value

For print.ppx and plot.ppx the value is NULL. For unitname.ppx the value is an object of class "units". For unitname<-.ppx and scale.ppx the value is another object of class "ppx".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

ppx, unitname

[Package spatstat.geom version 3.3-5 Index]