multiplicity.ppp {spatstat.geom}	R Documentation

Count Multiplicity of Duplicate Points

Description

Counts the number of duplicates for each point in a spatial point pattern.

Usage

 multiplicity(x)

 ## S3 method for class 'ppp'
multiplicity(x)

 ## S3 method for class 'ppx'
multiplicity(x)

 ## S3 method for class 'data.frame'
multiplicity(x)

 ## Default S3 method:
multiplicity(x)

Arguments

x
A spatial point pattern (object of class "ppp" or "ppx") or a vector, matrix or data frame.

Details

Two points in a point pattern are deemed to be identical if their x,y coordinates are the same, and their marks are also the same (if they carry marks). The Examples section illustrates how it is possible for a point pattern to contain a pair of identical points.

For each point in x, the function multiplicity counts how many points are identical to it, and returns the vector of counts.

The argument x can also be a vector, a matrix or a data frame. When x is a vector, m <- multiplicity(x) is a vector of the same length as x, and m[i] is the number of elements of x that are identical to x[i]. When x is a matrix or data frame, m <- multiplicity(x) is a vector of length equal to the number of rows of x, and m[i] is the number of rows of x that are identical to the ith row.

Value

A vector of integers (multiplicities) of length equal to the number of points in x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

, Rolf Turner rolfturner@posteo.net

and Sebastian Meyer.

See Also

ppp.object, duplicated.ppp, unique.ppp

Examples

   X <- ppp(c(1,1,0.5,1), c(2,2,1,2), window=square(3), check=FALSE)
   m <- multiplicity(X)

   # unique points in X, marked by their multiplicity
   first <- !duplicated(X)
   Y <- X[first] %mark% m[first]

[Package spatstat.geom version 3.3-5 Index]