nestsplit {spatstat.geom}	R Documentation

Nested Split

Description

Applies two splitting operations to a point pattern, producing a list of lists of patterns.

Usage

  nestsplit(X, ...)

Arguments

X
Point pattern to be split. Object of class "ppp".

...
Data determining the splitting factors or splitting regions. See Details.

Details

This function splits the point pattern X into several sub-patterns using split.ppp, then splits each of the sub-patterns into sub-sub-patterns using split.ppp again. The result is a hyperframe containing the sub-sub-patterns and two factors indicating the grouping.

The arguments ... determine the two splitting factors or splitting regions. Each argument may be:

a factor (of length equal to the number of points in X)

the name of a column of marks of X (provided this column contains factor values)

a tessellation (class "tess")

a pixel image (class "im") with factor values

a window (class "owin")

identified by name (in the form name=value) as one of the formal arguments of quadrats or tess

The arguments will be processed to yield a list of two splitting factors/tessellations. The splits will be applied to X consecutively to produce the sub-sub-patterns.

Value

A hyperframe with three columns. The first column contains the sub-sub-patterns. The second and third columns are factors which identify the grouping according to the two splitting factors.

Author(s)

Original idea by Ute Hahn. Code by Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

split.ppp, quantess

Examples

  # factor and tessellation
  Nft <- nestsplit(amacrine, marks(amacrine), quadrats(amacrine, 3, 1))
  Ntf <- nestsplit(amacrine, quadrats(amacrine, 3, 1), marks(amacrine))
  Ntf

  # two factors
  big <- with(marks(betacells), area > 300)
  Nff <- nestsplit(betacells, "type", factor(big))

  # two tessellations
  Tx <- quantess(redwood, "x", 4)
  Td <- dirichlet(runifrect(5, Window(redwood)))
  Ntt <- nestsplit(redwood, Td, Tx)
  Ntt2 <- nestsplit(redwood, Td, ny=3)

[Package spatstat.geom version 3.3-5 Index]