nobjects {spatstat.geom}	R Documentation

Count Number of Geometrical Objects in a Spatial Dataset

Description

A generic function to count the number of geometrical objects in a spatial dataset.

Usage

  nobjects(x)

  ## S3 method for class 'ppp'
nobjects(x)

  ## S3 method for class 'ppx'
nobjects(x)

  ## S3 method for class 'psp'
nobjects(x)

  ## S3 method for class 'tess'
nobjects(x)

Arguments

x
A dataset.

Details

The generic function nobjects counts the number of geometrical objects in the spatial dataset x.

The methods for point patterns (classes "ppp" and "ppx", embracing "pp3" and "lpp") count the number of points in the pattern.

The method for line segment patterns (class "psp") counts the number of line segments in the pattern.

The method for tessellations (class "tess") counts the number of tiles of the tessellation.

Value

A single integer.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

npoints

Examples

   nobjects(redwood)
   nobjects(edges(letterR))
   nobjects(dirichlet(cells))

[Package spatstat.geom version 3.3-5 Index]