npoints {spatstat.geom}	R Documentation

Number of Points in a Point Pattern

Description

Returns the number of points in a point pattern of any kind.

Usage

  npoints(x)
  ## S3 method for class 'ppp'
npoints(x)
  ## S3 method for class 'pp3'
npoints(x)
  ## S3 method for class 'ppx'
npoints(x)

Arguments

x
A point pattern (object of class "ppp", "pp3", "ppx" or some other suitable class).

Details

This function returns the number of points in a point pattern. The function npoints is generic with methods for the classes "ppp", "pp3", "ppx" and possibly other classes.

Value

Integer.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

ppp.object, print.pp3, print.ppx.

Examples

   npoints(cells)

[Package spatstat.geom version 3.3-5 Index]