nsegments {spatstat.geom}	R Documentation

Number of Line Segments in a Line Segment Pattern

Description

Returns the number of line segments in a line segment pattern.

Usage

   nsegments(x)

   ## S3 method for class 'psp'
nsegments(x)

Arguments

x
A line segment pattern, i.e. an object of class psp, or an object containing a linear network.

Details

This function is generic, with methods for classes psp, linnet and lpp.

Value

Integer.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

npoints(), psp.object()

Examples

nsegments(copper$Lines)
nsegments(copper$SouthLines)

[Package spatstat.geom version 3.3-5 Index]