nvertices {spatstat.geom}	R Documentation

Count Number of Vertices

Description

Count the number of vertices in an object for which vertices are well-defined.

Usage

  nvertices(x, ...)

  ## S3 method for class 'owin'
nvertices(x, ...)

  ## Default S3 method:
nvertices(x, ...)

Arguments

x
A window (object of class "owin"), or some other object which has vertices.

...
Currently ignored.

Details

This function counts the number of vertices of x as they would be returned by vertices(x). It is more efficient than executing npoints(vertices(x)).

Value

A single integer.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk and Suman Rakshit.

See Also

vertices

Examples

  nvertices(square(2))
  nvertices(letterR)

[Package spatstat.geom version 3.3-5 Index]