overlap.owin {spatstat.geom}	R Documentation

Compute Area of Overlap

Description

Computes the area of the overlap (intersection) of two windows.

Usage

overlap.owin(A, B)

Arguments

A,B
Windows (objects of class "owin").

Details

This function computes the area of the overlap between the two windows A and B.

If one of the windows is a binary mask, then both windows are converted to masks on the same grid, and the area is computed by counting pixels. Otherwise, the area is computed analytically (using the discrete Stokes theorem).

Value

A single numeric value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

intersect.owin, area.owin, setcov.

Examples

  A <- square(1)
  B <- shift(A, c(0.3, 0.2))
  overlap.owin(A, B)

[Package spatstat.geom version 3.3-5 Index]