pairdist.psp {spatstat.geom}	R Documentation

Pairwise distances between line segments

Description

Computes the matrix of distances between all pairs of line segments in a line segment pattern.

Usage

  ## S3 method for class 'psp'
pairdist(X, ..., method="C", type="Hausdorff")

Arguments

X
A line segment pattern (object of class "psp").

...
Ignored.

method
String specifying which method of calculation to use. Values are "C" and "interpreted". Usually not specified.

type
Type of distance to be computed. Options are "Hausdorff" and "separation". Partial matching is used.

Details

This function computes the distance between each pair of line segments in X, and returns the matrix of distances.

This is a method for the generic function pairdist for the class "psp".

The distances between line segments are measured in one of two ways:

if type="Hausdorff", distances are computed in the Hausdorff metric. The Hausdorff distance between two line segments is the maximum distance from any point on one of the segments to the nearest point on the other segment.

if type="separation", distances are computed as the minimum distance from a point on one line segment to a point on the other line segment. For example, line segments which cross over each other have separation zero.

The argument method is not normally used. It is retained only for checking the validity of the software. If method = "interpreted" then the distances are computed using interpreted R code only. If method="C" (the default) then compiled C code is used, which is somewhat faster.

Value

A square matrix whose [i,j] entry is the distance between the line segments numbered i and j.

Distance values

The values returned by pairdist(X) are distances, expressed as multiples of the unit of length of the spatial coordinates in X. The unit of length is given by unitname(X).

Note that, if the unit of length in X is a composite expression such as ‘2 microns’, then the values of pairdist(X) are expressed as multiples of 2 microns, rather than being expressed in microns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

crossdist, nndist, pairdist.ppp

Examples

   L <- psp(runif(10), runif(10), runif(10), runif(10), owin())
   D <- pairdist(L)
   S <- pairdist(L, type="sep")

[Package spatstat.geom version 3.3-5 Index]