perimeter {spatstat.geom}	R Documentation

Perimeter Length of Window

Description

Computes the perimeter length of a window

Usage

perimeter(w)

Arguments

w
A window (object of class "owin") or data that can be converted to a window by as.owin.

Details

This function computes the perimeter (length of the boundary) of the window w. If w is a rectangle or a polygonal window, the perimeter is the sum of the lengths of the edges of w. If w is a mask, it is first converted to a polygonal window using as.polygonal, then staircase edges are removed using simplify.owin, and the perimeter of the resulting polygon is computed.

Value

A numeric value giving the perimeter length of the window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

area.owin diameter.owin, owin.object, as.owin

Examples

  perimeter(square(3))
  perimeter(letterR)
  if(interactive()) print(perimeter(as.mask(letterR)))

[Package spatstat.geom version 3.3-5 Index]