pixelcentres {spatstat.geom}	R Documentation

Extract Pixel Centres as Point Pattern

Description

Given a pixel image or binary mask window, extract the centres of all pixels and return them as a point pattern.

Usage

pixelcentres(X, W = NULL, ...)

Arguments

X
Pixel image (object of class "im") or window (object of class "owin").

W
Optional window to contain the resulting point pattern.

...
Optional arguments defining the pixel resolution.

Details

If the argument X is a pixel image, the result is a point pattern, consisting of the centre of every pixel whose pixel value is not NA.

If X is a window which is a binary mask, the result is a point pattern consisting of the centre of every pixel inside the window (i.e. every pixel for which the mask value is TRUE).

Otherwise, X is first converted to a window, then converted to a mask using as.mask, then handled as above.

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

, Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

raster.xy

Examples

  pixelcentres(letterR, dimyx=5)

[Package spatstat.geom version 3.3-5 Index]