pixellate {spatstat.geom}	R Documentation

Convert Spatial Object to Pixel Image

Description

Convert a spatial object to a pixel image by measuring the amount of stuff in each pixel.

Usage

pixellate(x, ...)

Arguments

x
Spatial object to be converted. A point pattern (object of class "ppp"), a window (object of class "owin"), a line segment pattern (object of class "psp"), or some other suitable data.

...
Arguments passed to methods.

Details

The function pixellate converts a geometrical object x into a pixel image, by measuring the amount of x that is inside each pixel.

If x is a point pattern, pixellate(x) counts the number of points of x falling in each pixel. If x is a window, pixellate(x) measures the area of intersection of each pixel with the window.

The function pixellate is generic, with methods for point patterns (pixellate.ppp), windows (pixellate.owin), and line segment patterns (pixellate.psp), See the separate documentation for these methods.

The related function as.im also converts x into a pixel image, but typically measures only the presence or absence of x inside each pixel.

Value

A pixel image (object of class "im").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

pixellate.ppp, pixellate.owin, pixellate.psp, as.im

[Package spatstat.geom version 3.3-5 Index]