plot.imlist {spatstat.geom}	R Documentation

Plot a List of Images

Description

Plots an array of pixel images.

Usage

  ## S3 method for class 'imlist'
plot(x, ..., plotcommand="image",
                          equal.ribbon=FALSE,
                          equal.scales=FALSE, 
                          ribmar=NULL)

  ## S3 method for class 'imlist'
image(x, ..., equal.ribbon=FALSE,
                                   equal.scales=FALSE, ribmar=NULL)

  ## S3 method for class 'listof'
image(x, ..., equal.ribbon=FALSE,
                                   equal.scales=FALSE, ribmar=NULL)

Arguments

x
An object of the class "imlist" representing a list of pixel images. Alternatively x may belong to the outdated class "listof".

...
Arguments passed to plot.solist to control the spatial arrangement of panels, and arguments passed to plot.im to control the display of each panel.

equal.ribbon
Logical. If TRUE, the colour maps of all the images will be the same. If FALSE, the colour map of each image is adjusted to the range of values of that image.

equal.scales
Logical. If TRUE, the images will be plotted using the same physical scale, and the plots will be aligned neatly where possible.

ribmar
Numeric vector of length 4 specifying the margins around the colour ribbon, if equal.ribbon=TRUE. Entries in the vector give the margin at the bottom, left, top, and right respectively, as a multiple of the height of a line of text.

plotcommand
Character string giving the name of a function to be used to display each image. Recognised by plot.imlist only.

Details

These are methods for the generic plot commands plot and image for the class "imlist". They are currently identical.

An object of class "imlist" represents a list of pixel images. (The outdated class "listof" is also handled.)

Each entry in the list x will be displayed as a pixel image, in an array of panels laid out on the same graphics display, using plot.solist. Individual panels are plotted by plot.im.

If equal.ribbon=FALSE (the default), the images are rendered using different colour maps, which are displayed as colour ribbons beside each image. If equal.ribbon=TRUE, the images are rendered using the same colour map, and a single colour ribbon will be displayed at the right side of the array. The colour maps and the placement of the colour ribbons are controlled by arguments ... passed to plot.im.

If equal.scales=TRUE, the images are plotted using the same physical scale, and the plots will be aligned neatly where possible. If equal.scales=FALSE (the default), images are plotted using equal amounts of space in the available plotting area, so they may be plotted at different physical scales.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

plot.solist, plot.im

Examples

 ## bei.extra is a list of pixel images on the same spatial domain
 Y <- solapply(bei.extra, scaletointerval)
 image(Y, equal.ribbon=TRUE, equal.scales=TRUE,
          main="",
          mar.panel=0, hsep=1,
          ribside="bottom",
          col.ticks="blue", col.axis="blue", cex.axis=1.2)

[Package spatstat.geom version 3.3-5 Index]