plot.onearrow {spatstat.geom}	R Documentation

Plot an Arrow

Description

Plots an object of class "onearrow".

Usage

 ## S3 method for class 'onearrow'
plot(x, ...,
  add = FALSE, main = "",
  retract = 0.05, headfraction = 0.25, headangle = 12, headnick = 0.1,
  col.head = NA, lwd.head = lwd, lwd = 1, col = 1,
  zap = FALSE, zapfraction = 0.07,
  pch = 1, cex = 1, do.plot = TRUE, do.points = FALSE, show.all = !add)

Arguments

x
Object of class "onearrow" to be plotted. This object is created by the command onearrow.

...
Additional graphics arguments passed to segments to control the appearance of the line.

add
Logical value indicating whether to add graphics to the existing plot (add=TRUE) or to start a new plot (add=FALSE).

main
Main title for the plot.

retract
Fraction of length of arrow to remove at each end.

headfraction
Length of arrow head as a fraction of overall length of arrow.

headangle
Angle (in degrees) between the outer edge of the arrow head and the shaft of the arrow.

headnick
Size of the nick in the trailing edge of the arrow head as a fraction of length of arrow head.

col.head,lwd.head
Colour and line style of the filled arrow head.

col,lwd
Colour and line style of the arrow shaft.

zap
Logical value indicating whether the arrow should include a Z-shaped (lightning-bolt) feature in the middle of the shaft.

zapfraction
Size of Z-shaped deviation as a fraction of total arrow length.

pch,cex
Plot character and character size for the two end points of the arrow, if do.points=TRUE.

do.plot
Logical. Whether to actually perform the plot.

do.points
Logical. Whether to display the two end points of the arrow as well.

show.all
Internal use only.

Details

The argument x should be an object of class "onearrow" created by the command onearrow.

Value

A window (class "owin") enclosing the plotted graphics.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

onearrow, yardstick

Examples

  oa <- onearrow(cells[c(1, 42)])
  oa
  plot(oa)
  plot(oa, zap=TRUE, do.points=TRUE, col.head="pink", col="red")

[Package spatstat.geom version 3.3-5 Index]