plot.pp3 {spatstat.geom}	R Documentation

Plot a Three-Dimensional Point Pattern

Description

Plots a three-dimensional point pattern.

Usage

 ## S3 method for class 'pp3'
plot(x, ..., eye=NULL, org=NULL, theta=25, phi=15,
                    type=c("p", "n", "h"),
                    box.back=list(col="pink"),
                    box.front=list(col="blue", lwd=2))

Arguments

x
Three-dimensional point pattern (object of class "pp3").

...
Arguments passed to points controlling the appearance of the points.

eye
Optional. Eye position. A numeric vector of length 3 giving the location from which the scene is viewed.

org
Optional. Origin (centre) of the view. A numeric vector of length 3 which will be at the centre of the view.

theta,phi
Optional angular coordinates (in degrees) specifying the direction from which the scene is viewed: theta is the azimuth and phi is the colatitude. Ignored if eye is given.

type
Type of plot: type="p" for points, type="h" for points on vertical lines, type="n" for box only.

box.front,box.back
How to plot the three-dimensional box that contains the points. A list of graphical arguments passed to segments, or a logical value indicating whether or not to plot the relevant part of the box. See Details.

Details

This is the plot method for objects of class "pp3". It generates a two-dimensional plot of the point pattern x and its containing box as if they had been viewed from the location specified by eye (or from the direction specified by theta and phi).

The edges of the box at the ‘back’ of the scene (as viewed from the eye position) are plotted first. Then the points are added. Finally the remaining ‘front’ edges are plotted. The arguments box.back and box.front specify graphical parameters for drawing the back and front edges, respectively. Alternatively box.back=FALSE specifies that the back edges shall not be drawn.

Note that default values of arguments to plot.pp3 can be set by spatstat.options("par.pp3").

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

pp3, spatstat.options.

Examples

  X <- osteo$pts[[1]]
  plot(X, main="Osteocyte lacunae, animal 1, brick 1",
       cex=1.5, pch=16)
  plot(X, type="h", main="", box.back=list(lty=3))

[Package spatstat.geom version 3.3-5 Index]