plot.quad {spatstat.geom}	R Documentation

Plot a Spatial Quadrature Scheme

Description

Plot a two-dimensional spatial quadrature scheme.

Usage

  ## S3 method for class 'quad'
plot(x, ..., main, add=FALSE, dum=list(), tiles=FALSE)

Arguments

x
The spatial quadrature scheme to be plotted. An object of class "quad".

...
extra arguments controlling the plotting of the data points of the quadrature scheme.

main
text to be displayed as a title above the plot.

add
Logical value indicating whether the graphics should be added to the current plot if there is one (add=TRUE) or whether a new plot should be initialised (add=FALSE, the default).

dum
list of extra arguments controlling the plotting of the dummy points of the quadrature scheme. See below.

tiles
Logical value indicating whether to display the tiles used to compute the quadrature weights.

Details

This is the plot method for quadrature schemes (objects of class "quad", see quad.object).

First the data points of the quadrature scheme are plotted (in their observation window) using plot.ppp with any arguments specified in ...

Then the dummy points of the quadrature scheme are plotted using plot.ppp with any arguments specified in dum.

By default the dummy points are superimposed onto the plot of data points. This can be overridden by including the argument add=FALSE in the list dum as shown in the examples. In this case the data and dummy point patterns are plotted separately.

See par and plot.ppp for other possible arguments controlling the plots.

Value

NULL.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

quad.object, plot.ppp, par

Examples

   Q <- quadscheme(nztrees)

   plot(Q, main="NZ trees: quadrature scheme")

   oldpar <- par(mfrow=c(2,1))
   plot(Q, main="NZ trees", dum=list(add=FALSE))
   par(oldpar)

[Package spatstat.geom version 3.3-5 Index]