plot.splitppp {spatstat.geom}	R Documentation

Plot a List of Point Patterns

Description

Plots a list of point patterns.

Usage

  ## S3 method for class 'splitppp'
plot(x, ..., main)

Arguments

x
A named list of point patterns, typically obtained from split.ppp.

...
Arguments passed to plot.listof which control the layout of the plot panels, their appearance, and the plot behaviour in individual plot panels.

main
Optional main title for the plot.

Details

This is the plot method for the class "splitppp". It is typically used to plot the result of the function split.ppp.

The argument x should be a named list of point patterns (objects of class "ppp", see ppp.object). Each of these point patterns will be plotted in turn using plot.ppp.

Plotting is performed by plot.listof.

Value

Null.

Error messages

If the error message ‘Figure margins too large’ occurs, ensure that equal.scales=FALSE and increase the values of mar.panel.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

plot.listof for arguments controlling the plot.

split.ppp, plot.ppp, ppp.object.

Examples

# Multitype point pattern
 plot(split(amacrine))
 plot(split(amacrine), main="", 
     panel.begin=function(i, y, ...) { plot(distmap(y), ribbon=FALSE, ...) })

[Package spatstat.geom version 3.3-5 Index]