plot.textstring {spatstat.geom}	R Documentation

Plot a Text String

Description

Plots an object of class "textstring".

Usage

## S3 method for class 'textstring'
plot(x, ..., do.plot = TRUE)

Arguments

x
Object of class "textstring" to be plotted. This object is created by the command textstring.

...
Additional graphics arguments passed to text to control the plotting of text.

do.plot
Logical value indicating whether to actually plot the text.

Details

The argument x should be an object of class "textstring" created by the command textstring.

This function displays the text using text.

Value

A window (class "owin") enclosing the plotted graphics.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

onearrow, yardstick

Examples

  W <- Window(humberside)
  te <- textstring(centroid.owin(W), txt="Humberside", cex=2.5)
  te
  plot(layered(W, te), main="")

[Package spatstat.geom version 3.3-5 Index]