polartess {spatstat.geom}	R Documentation

Tessellation Using Polar Coordinates

Description

Create a tessellation with tiles defined by polar coordinates (radius and angle).

Usage

polartess(W, ..., nradial = NULL, nangular = NULL,
          radii = NULL, angles = NULL,
          origin = NULL, sep = "x")

Arguments

W
A window (object of class "owin") or anything that can be coerced to a window using as.owin, such as a point pattern.

...
Ignored.

nradial
Number of tiles in the radial direction. A single integer. Ignored if radii is given.

nangular
Number of tiles in the angular coordinate. A single integer. Ignored if angles is given.

radii
The numeric values of the radii, defining the tiles in the radial direction. A numeric vector, of length at least 2, containing nonnegative numbers in increasing order. The value Inf is permitted.

angles
The numeric values of the angles defining the tiles in the angular coordinate. A numeric vector, of length at least 2, in increasing order, containing angles in radians.

origin
Location to be used as the origin of the polar coordinates. Either a numeric vector of length 2 giving the spatial location of the origin, or one of the strings "centroid", "midpoint", "left", "right", "top", "bottom", "topleft", "bottomleft", "topright" or "bottomright" indicating the location in the window.

sep
Argument passed to intersect.tess specifying the character string to be used as a separator when forming the names of the tiles.

Details

A tessellation will be formed from tiles defined by intervals in the polar coordinates r (radial distance from the origin) or θ (angle from the horizontal axis) or both. These tiles look like the cells on a dartboard.

If the argument radii is given, tiles will be demarcated by circles centred at the origin, with the specified radii. If radii is absent but nradial is given, then radii will default to a sequence of nradial+1 radii equally spaced from zero to the maximum possible radius. If neither radii nor nradial are given, the tessellation will not include circular arc boundaries.

If the argument angles is given, tiles will be demarcated by lines emanating from the origin at the specified angles. The angular values can be any real numbers; they will be interpreted as angles in radians modulo 2*pi, but they must be an increasing sequence of numbers. If angles is absent but nangular is given, then angles will default to a sequence of nangular+1 angles equally spaced from 0 to 2*pi. If neither angles nor nangular are given, the tessellation will not include linear boundaries.

Value

A tessellation (object of class "tess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

intersect.tess

To construct other kinds of tessellations, see tess, quadrats, hextess, venn.tess, dirichlet, delaunay, quantess, bufftess and rpoislinetess.

Examples

  Y <- c(2.8, 1.5)
  plot(polartess(letterR, nangular=6, radii=(0:4)/2, origin=Y),
       do.col=TRUE)

[Package spatstat.geom version 3.3-5 Index]