pp3 {spatstat.geom}	R Documentation

Three Dimensional Point Pattern

Description

Create a three-dimensional point pattern

Usage

pp3(x, y, z, ..., marks=NULL)

Arguments

x,y,z
Numeric vectors of equal length, containing Cartesian coordinates of points in three-dimensional space.

...
Arguments passed to as.box3 to determine the three-dimensional box in which the points have been observed.

marks
Optional. Vector, data frame, or hyperframe of mark values associated with the points.

Details

An object of class "pp3" represents a pattern of points in three-dimensional space. The points are assumed to have been observed by exhaustively inspecting a three-dimensional rectangular box. The boundaries of the box are included as part of the dataset.

Value

Object of class "pp3" representing a three dimensional point pattern. Also belongs to class "ppx".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

box3, print.pp3, ppx

Examples

   X <- pp3(runif(10), runif(10), runif(10),
            box3(c(0,1)),
            marks=rnorm(10))
   X

[Package spatstat.geom version 3.3-5 Index]