print.owin {spatstat.geom}	R Documentation

Print Brief Details of a Spatial Window

Description

Prints a very brief description of a window object.

Usage

  ## S3 method for class 'owin'
print(x, ..., prefix="window: ")

Arguments

x
Window (object of class "owin").

...
Ignored.

prefix
Character string to be printed at the start of the output.

Details

A very brief description of the window x is printed.

This is a method for the generic function print.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

print, print.ppp, summary.owin

Examples

  owin()  # the unit square

  W <- Window(demopat)
  W                    # just says it is polygonal
  as.mask(W)           # just says it is a binary image 

[Package spatstat.geom version 3.3-5 Index]