print.psp {spatstat.geom}	R Documentation

Print Brief Details of a Line Segment Pattern Dataset

Description

Prints a very brief description of a line segment pattern dataset.

Usage

  ## S3 method for class 'psp'
print(x, ...)

Arguments

x
Line segment pattern (object of class "psp").

...
Ignored.

Details

A very brief description of the line segment pattern x is printed.

This is a method for the generic function print.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

print, print.owin, summary.psp

Examples

  a <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  a

[Package spatstat.geom version 3.3-5 Index]