psp2mask {spatstat.geom}	R Documentation

Convert Line Segment Pattern to Binary Pixel Mask

Description

Converts a line segment pattern to a binary pixel mask by determining which pixels intersect the lines.

Usage

psp2mask(x, W=NULL, ...)
as.mask.psp(x, W=NULL, ...)

Arguments

x
Line segment pattern (object of class "psp").

W
Optional window (object of class "owin") determining the pixel raster.

...
Optional extra arguments passed to as.mask to determine the pixel resolution.

Details

The functions psp2mask and as.mask.psp are currently identical. In future versions of the package, as.mask.psp will be deprecated, and then removed.

This function converts a line segment pattern to a binary pixel mask by determining which pixels intersect the lines.

The pixel raster is determined by W and the optional arguments .... If W is missing or NULL, it defaults to the window containing x. Then W is converted to a binary pixel mask using as.mask. The arguments ... are passed to as.mask to control the pixel resolution.

Value

A window (object of class "owin") which is a binary pixel mask (type "mask").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

pixellate.psp, as.mask.

Use pixellate.psp if you want to measure the length of line in each pixel.

Examples

  X <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  plot(psp2mask(X))
  plot(X, add=TRUE, col="red")

[Package spatstat.geom version 3.3-5 Index]