rQuasi {spatstat.geom}	R Documentation

Generate Quasirandom Point Pattern in Given Window

Description

Generates a quasirandom pattern of points in any two-dimensional window.

Usage

rQuasi(n, W, type = c("Halton", "Hammersley"), ...)

Arguments

n
Maximum number of points to be generated.

W
Window (object of class "owin") in which to generate the points.

type
String identifying the quasirandom generator.

...
Arguments passed to the quasirandom generator.

Details

This function generates a quasirandom point pattern, using the quasirandom sequence generator Halton or Hammersley as specified.

If W is a rectangle, exactly n points will be generated.

If W is not a rectangle, n points will be generated in the containing rectangle as.rectangle(W), and only the points lying inside W will be retained.

Value

Point pattern (object of class "ppp") inside the window W.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

, Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk.

See Also

Halton

Examples

   plot(rQuasi(256, letterR))

[Package spatstat.geom version 3.3-5 Index]