rectdistmap {spatstat.geom}	R Documentation

Distance Map Using Rectangular Distance Metric

Description

Computes the distance map of a spatial region based on the rectangular distance metric.

Usage

rectdistmap(X, asp = 1, npasses=1, verbose=FALSE)

Arguments

X
A window (object of class "owin").

asp
Aspect ratio for the metric. See Details.

npasses
Experimental.

verbose
Logical value indicating whether to print trace information.

Details

This function computes the distance map of the spatial region X using the rectangular distance metric with aspect ratio asp. This metric is defined so that the set of all points lying at most 1 unit away from the origin (according to the metric) form a rectangle of width 1 and height asp.

Value

A pixel image (object of class "im").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

distmap

Examples

  V <- letterR
  Frame(V) <- grow.rectangle(Frame(V), 0.5)
  plot(rectdistmap(V))

[Package spatstat.geom version 3.3-5 Index]