reflect {spatstat.geom}	R Documentation

Reflect In Origin

Description

Reflects a geometrical object through the origin.

Usage

  reflect(X)

  ## S3 method for class 'im'
reflect(X)

  ## Default S3 method:
reflect(X)

Arguments

X
Any suitable dataset representing a two-dimensional object, such as a point pattern (object of class "ppp"), or a window (object of class "owin").

Details

The object X is reflected through the origin. That is, each point in X with coordinates (x,y) is mapped to the position (-x, -y).

This is equivalent to applying the affine transformation with matrix diag(c(-1,-1)). It is also equivalent to rotation about the origin by 180 degrees.

The command reflect is generic, with a method for pixel images and a default method.

Value

Another object of the same type, representing the result of reflection.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

affine, flipxy

Examples

  plot(reflect(as.im(letterR)))
  plot(reflect(letterR), add=TRUE)

[Package spatstat.geom version 3.3-5 Index]