relevel.im {spatstat.geom}	R Documentation

Reorder Levels of a Factor-Valued Image or Pattern

Description

For a pixel image with factor values, or a point pattern with factor-valued marks, the levels of the factor are re-ordered so that the level ref is first and the others are moved down.

Usage

  ## S3 method for class 'im'
relevel(x, ref, ...)

  ## S3 method for class 'ppp'
relevel(x, ref, ...)

  ## S3 method for class 'ppx'
relevel(x, ref, ...)

Arguments

x
A pixel image (object of class "im") with factor values, or a point pattern (object of class "ppp", "ppx", "lpp" or "pp3") with factor-valued marks.

ref
The reference level.

...
Ignored.

Details

These functions are methods for the generic relevel.

If x is a pixel image (object of class "im") with factor values, or a point pattern (object of class "ppp", "ppx", "lpp" or "pp3") with factor-valued marks, the levels of the factor are changed so that the level specified by ref comes first.

Value

Object of the same kind as x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

mergeLevels

Examples

  amacrine
  relevel(amacrine, "on")

[Package spatstat.geom version 3.3-5 Index]