rescale.im {spatstat.geom}	R Documentation

Convert Pixel Image to Another Unit of Length

Description

Converts a pixel image to another unit of length.

Usage

 ## S3 method for class 'im'
rescale(X, s, unitname)

Arguments

X
Pixel image (object of class "im").

s
Conversion factor: the new units are s times the old units.

unitname
Optional. New name for the unit of length. See unitname.

Details

This is a method for the generic function rescale.

The spatial coordinates of the pixels in X will be re-expressed in terms of a new unit of length that is s times the current unit of length given in X. (Thus, the coordinate values are divided by s, while the unit value is multiplied by s).

If s is missing, then the coordinates will be re-expressed in ‘native’ units; for example if the current unit is equal to 0.1 metres, then the coordinates will be re-expressed in metres.

The result is a pixel image representing the same data but re-expressed in a different unit.

Pixel values are unchanged. This may not be what you intended!

Value

Another pixel image (of class "im"), containing the same pixel values, but with pixel coordinates expressed in the new units.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

im, rescale, unitname, eval.im

Examples

# Bramble Canes data: 1 unit = 9 metres
  bramblecanes
# distance transform
  Z <- distmap(bramblecanes)
# convert to metres
# first alter the pixel values
  Zm <- eval.im(9 * Z)
# now rescale the pixel coordinates
  Z <- rescale(Zm, 1/9)
# or equivalently
  Z <- rescale(Zm)

[Package spatstat.geom version 3.3-5 Index]