rescale.ppp {spatstat.geom}	R Documentation

Convert Point Pattern to Another Unit of Length

Description

Converts a point pattern dataset to another unit of length.

Usage

 ## S3 method for class 'ppp'
rescale(X, s, unitname)

Arguments

X
Point pattern (object of class "ppp").

s
Conversion factor: the new units are s times the old units.

unitname
Optional. New name for the unit of length. See unitname.

Details

This is a method for the generic function rescale.

The spatial coordinates in the point pattern X (and its window) will be re-expressed in terms of a new unit of length that is s times the current unit of length given in X. (Thus, the coordinate values are divided by s, while the unit value is multiplied by s).

The result is a point pattern representing the same data but re-expressed in a different unit.

Mark values are unchanged.

If s is missing, then the coordinates will be re-expressed in ‘native’ units; for example if the current unit is equal to 0.1 metres, then the coordinates will be re-expressed in metres.

Value

Another point pattern (of class "ppp"), representing the same data, but expressed in the new units.

Note

The result of this operation is equivalent to the original point pattern. If you want to actually change the coordinates by a linear transformation, producing a point pattern that is not equivalent to the original one, use affine.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

unitname, rescale, rescale.owin, affine, rotate, shift

Examples

# Bramble Canes data: 1 unit = 9 metres
# convert to metres
  bram <- rescale(bramblecanes, 1/9)
# or equivalently
  bram <- rescale(bramblecanes)

[Package spatstat.geom version 3.3-5 Index]