ripras {spatstat.geom}	R Documentation

Estimate window from points alone

Description

Given an observed pattern of points, computes the Ripley-Rasson estimate of the spatial domain from which they came.

Usage

 ripras(x, y=NULL, shape="convex", f)

Arguments

x
vector of x coordinates of observed points, or a 2-column matrix giving x,y coordinates, or a list with components x,y giving coordinates (such as a point pattern object of class "ppp".)

y
(optional) vector of y coordinates of observed points, if x is a vector.

shape
String indicating the type of window to be estimated: either "convex" or "rectangle".

f
(optional) scaling factor. See Details.

Details

Given an observed pattern of points with coordinates given by x and y, this function computes an estimate due to Ripley and Rasson (1977) of the spatial domain from which the points came.

The points are assumed to have been generated independently and uniformly distributed inside an unknown domain D.

If shape="convex" (the default), the domain D is assumed to be a convex set. The maximum likelihood estimate of D is the convex hull of the points (computed by convexhull.xy). Analogously to the problems of estimating the endpoint of a uniform distribution, the MLE is not optimal. Ripley and Rasson's estimator is a rescaled copy of the convex hull, centred at the centroid of the convex hull. The scaling factor is 1/√{1 - \frac m n} where n is the number of data points and m the number of vertices of the convex hull. The scaling factor may be overridden using the argument f.

If shape="rectangle", the domain D is assumed to be a rectangle with sides parallel to the coordinate axes. The maximum likelihood estimate of D is the bounding box of the points (computed by bounding.box.xy). The Ripley-Rasson estimator is a rescaled copy of the bounding box, with scaling factor (n+1)/(n-1) where n is the number of data points, centred at the centroid of the bounding box. The scaling factor may be overridden using the argument f.

Value

A window (an object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

References

Ripley, B.D. and Rasson, J.-P. (1977) Finding the edge of a Poisson forest. Journal of Applied Probability, 14, 483 – 491.

See Also

owin, as.owin, bounding.box.xy, convexhull.xy

Examples

  x <- runif(30)
  y <- runif(30)
  w <- ripras(x,y)
  plot(owin(), main="ripras(x,y)")
  plot(w, add=TRUE)
  points(x,y)

  X <- runifrect(15)
  plot(X, main="ripras(X)")
  plot(ripras(X), add=TRUE)

  # two points insufficient
  ripras(c(0,1),c(0,0))
  # triangle
  ripras(c(0,1,0.5), c(0,0,1))
  # three collinear points
  ripras(c(0,0,0), c(0,1,2))

[Package spatstat.geom version 3.3-5 Index]