rlinegrid {spatstat.geom}	R Documentation

Generate grid of parallel lines with random displacement

Description

Generates a grid of parallel lines, equally spaced, inside the specified window.

Usage

  rlinegrid(angle = 45, spacing = 0.1, win = owin())

Arguments

angle
Common orientation of the lines, in degrees anticlockwise from the x axis.

spacing
Spacing between successive lines.

win
Window in which to generate the lines. An object of class "owin" or something acceptable to as.owin.

Details

The grid is randomly displaced from the origin.

Value

A line segment pattern (object of class "psp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

psp, rpoisline

Examples

  plot(rlinegrid(30, 0.05))

[Package spatstat.geom version 3.3-5 Index]