rotate {spatstat.geom}	R Documentation

Rotate

Description

Applies a rotation to any two-dimensional object, such as a point pattern or a window.

Usage

  rotate(X, ...)

Arguments

X
Any suitable dataset representing a two-dimensional object, such as a point pattern (object of class "ppp"), or a window (object of class "owin").

...
Data specifying the rotation.

Details

This is generic. Methods are provided for point patterns (rotate.ppp) and windows (rotate.owin).

Value

Another object of the same type, representing the result of rotating X through the specified angle.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

rotate.ppp, rotate.owin

[Package spatstat.geom version 3.3-5 Index]