rotate.ppp {spatstat.geom}	R Documentation

Rotate a Point Pattern

Description

Rotates a point pattern

Usage

 ## S3 method for class 'ppp'
rotate(X, angle=pi/2, ..., centre=NULL)

Arguments

X
A point pattern (object of class "ppp").

angle
Angle of rotation.

...
Arguments passed to rotate.owin affecting the handling of the observation window, if it is a binary pixel mask.

centre
Centre of rotation. Either a vector of length 2, or a character string (partially matched to "centroid", "midpoint" or "bottomleft"). The default is the coordinate origin c(0,0).

Details

The points of the pattern, and the window of observation, are rotated about the origin by the angle specified. Angles are measured in radians, anticlockwise. The default is to rotate the pattern 90 degrees anticlockwise. If the points carry marks, these are preserved.

Value

Another object of class "ppp" representing the rotated point pattern.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

ppp.object, rotate.owin

Examples

  X <- rotate(cells, pi/3)
  # plot(X)

[Package spatstat.geom version 3.3-5 Index]