round.ppp {spatstat.geom}	R Documentation

Apply Numerical Rounding to Spatial Coordinates

Description

Apply numerical rounding to the spatial coordinates of a point pattern.

Usage

## S3 method for class 'ppp'
round(x, digits = 0, ...)

## S3 method for class 'pp3'
round(x, digits = 0, ...)

## S3 method for class 'ppx'
round(x, digits = 0, ...)

Arguments

x
A spatial point pattern in any dimension (object of class "ppp", "pp3" or "ppx").

digits
integer indicating the number of decimal places.

...
Additional arguments passed to the default method.

Details

These functions are methods for the generic function round. They apply numerical rounding to the spatial coordinates of the point pattern x.

Value

A point pattern object, of the same class as x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

rounding.ppp to determine whether numbers have been rounded.

round in the Base package.

Examples

  round(cells, 1)

[Package spatstat.geom version 3.3-5 Index]