rounding.ppp {spatstat.geom}	R Documentation

Detect Numerical Rounding

Description

Given a numeric vector, or an object containing numeric spatial coordinates, determine whether the values have been rounded to a certain number of decimal places.

Usage

## S3 method for class 'ppp'
rounding(x)

## S3 method for class 'pp3'
rounding(x)

## S3 method for class 'ppx'
rounding(x)

Arguments

x
A point pattern (object of class ppp, pp3 or ppx).

Details

The functions documented here are methods for the generic rounding. They determine whether the coordinates of a spatial object have been rounded to a certain number of decimal places.

If the coordinates of the points in x are not all integers, then rounding(x) returns the smallest number of digits d after the decimal point such that round(coords(x), digits=d) is identical to coords(x). For example if rounding(x) = 2 then the coordinates of the points in x appear to have been rounded to 2 decimal places, and are multiples of 0.01.

If all the coordinates of the points in x are integers, then rounding(x) returns -d, where d is the smallest number of digits before the decimal point such that round(coords(x), digits=-d) is identical to coords(x). For example if rounding(x) = -3 then the coordinates of all points in x are multiples of 1000. If rounding(x) = 0 then the entries of x are integers but not multiples of 10.

If all coordinates of points in x are equal to 0, a value of 0 is returned.

Value

An integer.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

round.ppp, rounding

Examples

   rounding(cells)

[Package spatstat.geom version 3.3-5 Index]