scaletointerval {spatstat.geom}	R Documentation

Rescale Data to Lie Between Specified Limits

Description

Rescales a dataset so that the values range exactly between the specified limits.

Usage

  scaletointerval(x, from=0, to=1, xrange=range(x))
  ## Default S3 method:
scaletointerval(x, from=0, to=1, xrange=range(x))
  ## S3 method for class 'im'
scaletointerval(x, from=0, to=1, xrange=range(x))

Arguments

x
Data to be rescaled.

from,to
Lower and upper endpoints of the interval to which the values of x should be rescaled.

xrange
Optional range of values of x that should be mapped to the new interval.

Details

These functions rescale a dataset x so that its values range exactly between the limits from and to.

The method for pixel images (objects of class "im") applies this scaling to the pixel values of x.

Rescaling cannot be performed if the values in x are not interpretable as numeric, or if the values in x are all equal.

Value

An object of the same type as x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

scale

Examples

  X <- as.im(function(x,y) {x+y+3}, unit.square())
  summary(X)
  Y <- scaletointerval(X)
  summary(Y)

[Package spatstat.geom version 3.3-5 Index]