selfcrossing.psp {spatstat.geom}	R Documentation

Crossing Points in a Line Segment Pattern

Description

Finds any crossing points between the line segments in a line segment pattern.

Usage

  selfcrossing.psp(A)

Arguments

A
Line segment pattern (object of class "psp").

Details

This function finds any crossing points between different line segments in the line segment pattern A.

A crossing point occurs whenever one of the line segments in A intersects another line segment in A, at a nonzero angle of intersection.

Value

Point pattern (object of class "ppp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

crossing.psp, psp.object, ppp.object.

Examples

  a <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  plot(a, col="green", main="selfcrossing.psp")
  P <- selfcrossing.psp(a)
  plot(P, add=TRUE, col="red")

[Package spatstat.geom version 3.3-5 Index]