sessionLibs {spatstat.geom}	R Documentation

Print Names and Version Numbers of Libraries Loaded

Description

Prints the names and version numbers of libraries currently loaded by the user.

Usage

  sessionLibs()

Details

This function prints a list of the libraries loaded by the user in the current session, giving just their name and version number. It obtains this information from sessionInfo.

This function is not needed in an interactive R session because the package startup messages will usually provide this information.

Its main use is in an Sweave script, where it is needed because the package startup messages are not printed.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

sessionInfo

Examples

  sessionLibs()

[Package spatstat.geom version 3.3-5 Index]