shift.owin {spatstat.geom}	R Documentation

Apply Vector Translation To Window

Description

Applies a vector shift to a window

Usage

 ## S3 method for class 'owin'
shift(X, vec=c(0,0), ..., origin=NULL)

Arguments

X
Window (object of class "owin").

vec
Vector of length 2 representing a translation.

...
Ignored

origin
Location that will be shifted to the origin. Either a numeric vector of length 2 giving the location, or a point pattern containing only one point, or a list with two entries named x and y, or one of the character strings "centroid", "midpoint", "left", "right", "top", "bottom", "topleft", "bottomleft", "topright" or "bottomright" (partially matched).

Details

The window is translated by the vector vec. This is a method for the generic function shift.

If origin is given, the argument vec will be ignored; instead the shift will be performed so that the specified geometric location is shifted to the coordinate origin (0,0). The argument origin should be either a numeric vector of length 2 giving the spatial coordinates of a location, or one of the character strings "centroid", "midpoint", "left", "right", "top", "bottom", "topleft", "bottomleft", "topright" or "bottomright" (partially matched). If origin="centroid" then the centroid of the window will be shifted to the origin. If origin="midpoint" then the centre of the bounding rectangle of the window will be shifted to the origin. If origin="bottomleft" then the bottom left corner of the bounding rectangle of the window will be shifted to the origin, and so on.

Value

Another window (of class "owin") representing the result of applying the vector shift.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

shift, shift.ppp, periodify, rotate, affine, centroid.owin

Examples

  W <- owin(c(0,1),c(0,1))
  X <- shift(W, c(2,3))
  # plot(W)
  # no discernible difference except coordinates are different
  shift(W, origin="top")

[Package spatstat.geom version 3.3-5 Index]