shift.ppx {spatstat.geom}	R Documentation

Apply Vector Translation To Box Or Point Pattern In Arbitrary Dimension

Description

Applies a vector shift to a box or point pattern in arbitrary dimension (object of class "boxx" or "ppx").

Usage

 ## S3 method for class 'boxx'
shift(X, vec= 0, ...)
 ## S3 method for class 'ppx'
shift(X, vec = 0, ..., spatial = TRUE, temporal = TRUE, local = TRUE)

Arguments

X
Box or point pattern in arbitrary dimension (object of class "boxx" or "ppx").

vec
Either a single numeric or a vector of the same length as the dimension of the spatial and/or temporal and/or local domain.

...
Ignored

spatial,temporal,local
Logical to indicate whether or not to shift this type of coordinates for the ppx method.

Details

This is a method for the generic function shift.

Value

For shift.boxx, another "boxx" object and for shift.ppx another "ppx" object. In both cases the new object represents the result of applying the vector shift.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

shift, boxx, ppx

Examples

  vec <- c(2,3)
  dom <- boxx(c(0,1), c(0,1))
  X <- ppx(coords(cells), domain = dom)
  shift(dom, vec)
  Xs <- shift(X, vec)
  Xs
  head(coords(X), n = 3)
  head(coords(Xs), n = 3)

[Package spatstat.geom version 3.3-5 Index]