sidelengths.owin {spatstat.geom}	R Documentation

Side Lengths of Enclosing Rectangle of a Window

Description

Computes the side lengths of the (enclosing rectangle of) a window.

Usage

 ## S3 method for class 'owin'
sidelengths(x)

 ## S3 method for class 'owin'
shortside(x)

Arguments

x
A window whose side lengths will be computed. Object of class "owin".

Details

The functions shortside and sidelengths are generic. The functions documented here are the methods for the class "owin".

sidelengths.owin computes the side-lengths of the enclosing rectangle of the window x.

For safety, both functions give a warning if the window is not a rectangle. To suppress the warning, first convert the window to a rectangle using as.rectangle.

shortside.owin computes the minimum of the two side-lengths.

Value

For sidelengths.owin, a numeric vector of length 2 giving the side-lengths (x then y) of the enclosing rectangle. For shortside.owin, a numeric value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

shortside, sidelengths for the generic functions.

area.owin, diameter.owin, perimeter for other geometric calculations on "owin" objects.

owin, as.owin.

Examples

  w <- owin(c(0,2),c(-1,3))
  sidelengths(w)
  shortside(as.rectangle(letterR))

[Package spatstat.geom version 3.3-5 Index]