solist {spatstat.geom}	R Documentation

List of Two-Dimensional Spatial Objects

Description

Make a list of two-dimensional spatial objects.

Usage

solist(..., check=TRUE, promote=TRUE, demote=FALSE, .NameBase)

Arguments

...
Any number of objects, each representing a two-dimensional spatial dataset.

check
Logical value. If TRUE, check that each of the objects is a 2D spatial object.

promote
Logical value. If TRUE, test whether all objects belong to the same class, and if so, promote the list of objects to the appropriate class of list.

demote
Logical value determining what should happen if any of the objects is not a 2D spatial object: if demote=FALSE (the default), a fatal error occurs; if demote=TRUE, a list of class "anylist" is returned.

.NameBase
Optional. Character string. If the ... arguments have no names, then the entries of the resulting list will be given names that start with .NameBase.

Details

This command creates an object of class "solist" (spatial object list) which represents a list of two-dimensional spatial datasets. The datasets do not necessarily belong to the same class.

Typically the intention is that the datasets in the list should be treated in the same way, for example, they should be plotted side-by-side. The spatstat package provides a plotting function, plot.solist, and many other functions for this class.

In the spatstat package, various functions produce an object of class "solist". For example, when a point pattern is split into several point patterns by split.ppp, or an image is split into several images by split.im, the result is of class "solist".

If check=TRUE then the code will check whether all objects in ... belong to the classes of two-dimensional spatial objects defined in the spatstat package. They do not have to belong to the same class. Set check=FALSE for efficiency, but only if you are sure that all the objects are valid.

If some of the objects in ... are not two-dimensional spatial objects, the action taken depends on the argument demote. If demote=TRUE, the result will belong to the more general class "anylist" instead of "solist". If demote=FALSE (the default), an error occurs.

If promote=TRUE then the code will check whether all the objects ... belong to the same class. If they are all point patterns (class "ppp"), the result will also belong to the class "ppplist". If they are all pixel images (class "im"), the result will also belong to the class "imlist".

Use as.solist to convert a list to a "solist".

Value

A list, usually belonging to the class "solist".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

as.solist, anylist, solapply

Examples

  solist(cells, distmap(cells), quadratcount(cells))
  solist(cells, japanesepines, redwood, .NameBase="Pattern")

[Package spatstat.geom version 3.3-5 Index]