spatstat.geom-deprecated {spatstat.geom}	R Documentation

Deprecated spatstat.geom functions

Description

Deprecated spatstat.geom functions.

Usage

## S3 method for class 'owin'
as.psp(x, ..., window=NULL, 
       check=spatstat.options("checksegments"), fatal=TRUE)
circumradius(x, ...)
## S3 method for class 'owin'
circumradius(x, ...)
## S3 method for class 'ppp'
circumradius(x, ...)

Details

These functions are deprecated, and will eventually be deleted from the spatstat.geom package.

as.psp.owin has been replaced by edges.

circumradius is replaced by the more appropriately named boundingradius.

Value

as.psp.owin returns a window (object of class "owin").

circumradius returns a numeric value.

[Package spatstat.geom version 3.3-5 Index]