square {spatstat.geom}	R Documentation

Square Window

Description

Creates a square window

Usage

 square(r=1, unitname=NULL)
 unit.square()

Arguments

r
Numeric. The side length of the square, or a vector giving the minimum and maximum coordinate values.

unitname
Optional. Name of unit of length. Either a single character string, or a vector of two character strings giving the singular and plural forms, respectively.

Details

If r is a number, square(r) is a shortcut for creating a window object representing the square [0,r] * [0,r]. It is equivalent to the command owin(c(0,r),c(0,r)).

If r is a vector of length 2, then square(r) creates the square with x and y coordinates ranging from r[1] to r[2].

unit.square creates the unit square [0,1] * [0,1]. It is equivalent to square(1) or square() or owin(c(0,1),c(0,1)).

These commands are included for convenience, and to improve the readability of some code.

Value

An object of class "owin" (see owin.object) specifying a window.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

owin.object, owin

Examples

 W <- square(10)
 W <- square(c(-1,1))

[Package spatstat.geom version 3.3-5 Index]