stratrand {spatstat.geom}	R Documentation

Stratified random point pattern

Description

Generates a “stratified random” pattern of points in a window, by dividing the window into rectangular tiles and placing k random points in each tile.

Usage

 stratrand(window, nx, ny, k = 1)

Arguments

window
A window. An object of class owin, or data in any format acceptable to as.owin().

nx
Number of tiles in each row.

ny
Number of tiles in each column.

k
Number of random points to generate in each tile.

Details

The bounding rectangle of window is divided into a regular nx * ny grid of rectangular tiles. In each tile, k random points are generated independently with a uniform distribution in that tile.

Note that some of these grid points may lie outside the window, if window is not of type "rectangle". The function inside.owin can be used to select those grid points which do lie inside the window. See the examples.

This function is useful in creating dummy points for quadrature schemes (see quadscheme) as well as in simulating random point patterns.

Value

A list with two components x and y, which are numeric vectors giving the coordinates of the random points.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

quad.object, quadscheme, inside.owin, gridcentres

Examples

  w <- unit.square()
  xy <- stratrand(w, 10, 10)
  # plot(w)
  # points(xy)

  # polygonal boundary
  bdry <- list(x=c(0.1,0.3,0.7,0.4,0.2),
               y=c(0.1,0.1,0.5,0.7,0.3))
  w <- owin(c(0,1), c(0,1), poly=bdry)
  xy <- stratrand(w, 10, 10, 3)
  # plot(w)
  # points(xy)

  # determine which grid points are inside polygon
  ok <- inside.owin(xy$x, xy$y, w)
  # plot(w)
  # points(xy$x[ok], xy$y[ok])

[Package spatstat.geom version 3.3-5 Index]