summary.distfun {spatstat.geom}	R Documentation

Summarizing a Function of Spatial Location

Description

Prints a useful summary of a function of spatial location.

Usage

  ## S3 method for class 'distfun'
summary(object, ...)

  ## S3 method for class 'funxy'
summary(object, ...)

Arguments

object
An object of class "distfun" or "funxy" representing a function of spatial coordinates.

...
Arguments passed to as.mask controlling the pixel resolution used to compute the summary.

Details

These are the summary methods for the classes "funxy" and "distfun".

An object of class "funxy" represents a function of spatial location, defined in a particular region of space. This includes objects of the special class "distfun" which represent distance functions.

The summary method computes a summary of the function values. The function is evaluated on a grid of locations using as.im and numerical values at these locations are summarised using summary.im. The pixel resolution for the grid of locations is determined by the arguments ... which are passed to as.mask.

Value

For summary.funxy the result is an object of class "summary.funxy". For summary.distfun the result is an object of class "summary.distfun". There are print methods for these classes.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

distfun, funxy

Examples

   f <- function(x,y) { x^2 + y^2 - 1} 
   g <- funxy(f, square(2))
   summary(g)

   summary(distfun(cells))
   summary(distfun(cells), dimyx=256)

[Package spatstat.geom version 3.3-5 Index]