summary.owin {spatstat.geom}	R Documentation

Summary of a Spatial Window

Description

Prints a useful description of a window object.

Usage

  ## S3 method for class 'owin'
summary(object, ...)

Arguments

object
Window (object of class "owin").

...
Ignored.

Details

A useful description of the window object is printed.

This is a method for the generic function summary.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

summary, summary.ppp, print.owin

Examples

  summary(owin())  # the unit square

  W <- Window(demopat)  # weird polygonal window
  summary(W)           # describes it

  summary(as.mask(W))  # demonstrates current pixel resolution

[Package spatstat.geom version 3.3-5 Index]