summary.ppp {spatstat.geom}	R Documentation

Summary of a Point Pattern Dataset

Description

Prints a useful summary of a point pattern dataset.

Usage

 ## S3 method for class 'ppp'
summary(object, ..., checkdup=TRUE)

Arguments

object
Point pattern (object of class "ppp").

...
Ignored.

checkdup
Logical value indicating whether to check for the presence of duplicate points.

Details

A useful summary of the point pattern object is printed.

This is a method for the generic function summary.

If checkdup=TRUE, the pattern will be checked for the presence of duplicate points, using duplicated.ppp. This can be time-consuming if the pattern contains many points, so the checking can be disabled by setting checkdup=FALSE.

If the point pattern was generated by simulation using rmh, the parameters of the algorithm are printed.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

summary, summary.owin, print.ppp

Examples

  summary(cells)  # plain vanilla point pattern

  # multitype point pattern
  woods <- lansing
  
  summary(woods) # tabulates frequencies of each mark
  
  # numeric marks
  trees <- longleaf
  
  summary(trees) # prints summary.default(marks(trees))

  # weird polygonal window
  summary(demopat)  # describes it

[Package spatstat.geom version 3.3-5 Index]