summary.psp {spatstat.geom}	R Documentation

Summary of a Line Segment Pattern Dataset

Description

Prints a useful summary of a line segment pattern dataset.

Usage

  ## S3 method for class 'psp'
summary(object, ...)

Arguments

object
Line segment pattern (object of class "psp").

...
Ignored.

Details

A useful summary of the line segment pattern object is printed.

This is a method for the generic function summary.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

summary, summary.owin, print.psp

Examples

  a <- psp(runif(10), runif(10), runif(10), runif(10), window=owin())
  summary(a)  # describes it

[Package spatstat.geom version 3.3-5 Index]