summary.solist {spatstat.geom}	R Documentation

Summary of a List of Spatial Objects

Description

Prints a useful summary of each entry in a list of two-dimensional spatial objects.

Usage

 ## S3 method for class 'solist'
summary(object, ...)

Arguments

object
An object of class "solist".

...
Ignored.

Details

This is a method for the generic function summary.

An object of the class "solist" is effectively a list of two-dimensional spatial datasets. See solist.

This function extracts a useful summary of each of the datasets.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

solist, summary, plot.solist

Examples

  x <- solist(cells, japanesepines, redwood)
  summary(x)

[Package spatstat.geom version 3.3-5 Index]