summary.splitppp {spatstat.geom}	R Documentation

Summary of a Split Point Pattern

Description

Prints a useful summary of a split point pattern.

Usage

 ## S3 method for class 'splitppp'
summary(object, ...)

Arguments

object
Split point pattern (object of class "splitppp", effectively a list of point patterns, usually created by split.ppp).

...
Ignored.

Details

This is a method for the generic function summary.

An object of the class "splitppp" is effectively a list of point patterns (objects of class "ppp") representing different sub-patterns of an original point pattern.

This function extracts a useful summary of each of the sub-patterns.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

summary, split, split.ppp

Examples

  summary(split(amacrine))

[Package spatstat.geom version 3.3-5 Index]