tiles {spatstat.geom}	R Documentation

Extract List of Tiles in a Tessellation

Description

Extracts a list of the tiles that make up a tessellation.

Usage

tiles(x)

Arguments

x
A tessellation (object of class "tess").

Details

A tessellation is a collection of disjoint spatial regions (called tiles) that fit together to form a larger spatial region. See tess.

The tiles that make up the tessellation x are returned in a list.

Value

A list of windows (objects of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

tess, tilenames, tile.areas, tiles.empty

Examples

  A <- tess(xgrid=0:2,ygrid=0:2)
  tiles(A)
  v <- as.im(function(x,y){factor(round(x^2 + y^2))}, W=owin())
  E <- tess(image=v)
  tiles(E)

[Package spatstat.geom version 3.3-5 Index]