tiles.empty {spatstat.geom}	R Documentation

Check For Empty Tiles in a Tessellation

Description

Checks whether each tile in a tessellation is empty or non-empty.

Usage

tiles.empty(x)

Arguments

x
A tessellation (object of class "tess").

Details

A tessellation is a collection of disjoint spatial regions (called tiles) that fit together to form a larger spatial region. See tess.

It is possible for some tiles of a tessellation to be empty. For example, this can happen when the tessellation x is obtained by restricting another tessellation y to a smaller spatial domain w.

The function tiles.empty checks whether each tile is empty or non-empty. The result is a logical vector, with entries equal to TRUE when the corresponding tile is empty. Results are given in the same order as the tiles would be listed by tiles(x).

Value

A logical vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

tess, tiles, tilenames, tile.areas

Examples

  A <- tess(xgrid=0:2,ygrid=0:2)
  tiles.empty(A)
  v <- as.im(function(x,y){factor(round(x^2 + y^2))}, W=owin())
  E <- tess(image=v)
  tiles.empty(E)

[Package spatstat.geom version 3.3-5 Index]