timeTaken {spatstat.geom}	R Documentation

Extract the Total Computation Time

Description

Given an object or objects that contain timing information (reporting the amount of computer time taken to compute each object), this function extracts the timing data and evaluates the total time taken.

Usage

timeTaken(..., warn=TRUE)

Arguments

...
One or more objects of class "timed" containing timing data.

warn
Logical value indicating whether a warning should be issued if some of the arguments do not contain timing information.

Details

An object of class "timed" contains information on the amount of computer time that was taken to compute the object. See timed.

This function extracts the timing information from one or more such objects, and calculates the total time.

Value

An object inheriting the class "timed".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

timed

Examples

  A <- timed(minnndist(bei))
  B <- timed(minnndist(redwood))
  A
  B
  timeTaken(A,B)

[Package spatstat.geom version 3.3-5 Index]