triangulate.owin {spatstat.geom}	R Documentation

Decompose Window into Triangles

Description

Given a spatial window, this function decomposes the window into disjoint triangles. The result is a tessellation of the window in which each tile is a triangle.

Usage

triangulate.owin(W)

Arguments

W
Window (object of class "owin").

Details

The window W will be decomposed into disjoint triangles. The result is a tessellation of W in which each tile is a triangle. All triangle vertices lie on the boundary of the original polygon.

The window is first converted to a polygonal window using as.polygonal. The vertices of the polygonal window are extracted, and the Delaunay triangulation of these vertices is computed using delaunay. Each Delaunay triangle is intersected with the window: if the result is not a triangle, the triangulation procedure is applied recursively to this smaller polygon.

Value

Tessellation (object of class "tess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

tess, delaunay, as.polygonal

Examples

  plot(triangulate.owin(letterR))

[Package spatstat.geom version 3.3-5 Index]