trim.rectangle {spatstat.geom}	R Documentation

Cut margins from rectangle

Description

Trims a margin from a rectangle.

Usage

 trim.rectangle(W, xmargin=0, ymargin=xmargin)

Arguments

W
A window (object of class "owin"). Must be of type "rectangle".

xmargin
Width of horizontal margin to be trimmed. A single nonnegative number, or a vector of length 2 indicating margins of unequal width at left and right.

ymargin
Height of vertical margin to be trimmed. A single nonnegative number, or a vector of length 2 indicating margins of unequal width at bottom and top.

Details

This is a simple convenience function to trim off a margin of specified width and height from each side of a rectangular window. Unequal margins can also be trimmed.

Value

Another object of class "owin" representing the window after margins are trimmed.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

grow.rectangle, erosion, owin.object

Examples

  w <- square(10)
  # trim a margin of width 1 from all four sides
  square9 <- trim.rectangle(w, 1)

  # trim margin of width 3 from the right side
  # and margin of height 4 from top edge.
  v <- trim.rectangle(w, c(0,3), c(0,4))

[Package spatstat.geom version 3.3-5 Index]