tweak.colourmap {spatstat.geom}	R Documentation

Change Colour Values in a Colour Map

Description

Assign new colour values to some of the entries in a colour map.

Usage

tweak.colourmap(m, col, ..., inputs=NULL, range=NULL)

Arguments

m
A colour map (object of class "colourmap").

inputs
Input values to the colour map, to be assigned new colours. Incompatible with range.

range
Numeric vector of length 2 specifying a range of numerical values which should be assigned a new colour. Incompatible with inputs.

col
Replacement colours for the specified inputs or the specified range of values.

...
Other arguments are ignored.

Details

This function changes the colour map m by assigning new colours to each of the input values specified by inputs, or by assigning a single new colour to the range of input values specified by range.

The modified colour map is returned.

Value

Another colour map (object of class "colourmap").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

colourmap, interp.colourmap, colouroutputs, colourtools.

Examples

  co <- colourmap(rainbow(32), range=c(0,1))
  plot(tweak.colourmap(co, inputs=c(0.5, 0.6), "white"))
  plot(tweak.colourmap(co, range=c(0.5,0.6), "white"))

[Package spatstat.geom version 3.3-5 Index]