unique.ppp {spatstat.geom}	R Documentation

Extract Unique Points from a Spatial Point Pattern

Description

Removes any points that are identical to other points in a spatial point pattern.

Usage

 ## S3 method for class 'ppp'
unique(x, ..., warn=FALSE)

 ## S3 method for class 'ppx'
unique(x, ..., warn=FALSE)

Arguments

x
A spatial point pattern (object of class "ppp" or "ppx").

...
Arguments passed to duplicated.ppp or duplicated.data.frame.

warn
Logical. If TRUE, issue a warning message if any duplicated points were found.

Details

These are methods for the generic function unique for point pattern datasets (of class "ppp", see ppp.object, or class "ppx").

This function removes duplicate points in x, and returns a point pattern.

Two points in a point pattern are deemed to be identical if their x,y coordinates are the same, and their marks are the same (if they carry marks). This is the default rule: see duplicated.ppp for other options.

Value

Another point pattern object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

ppp.object, duplicated.ppp, multiplicity.ppp

Examples

   X <- ppp(c(1,1,0.5), c(2,2,1), window=square(3))
   unique(X)
   unique(X, rule="deldir")

[Package spatstat.geom version 3.3-5 Index]