unmark {spatstat.geom}	R Documentation

Remove Marks

Description

Remove the mark information from a spatial dataset.

Usage

 unmark(X)
 ## S3 method for class 'ppp'
unmark(X)
 ## S3 method for class 'splitppp'
unmark(X)
 ## S3 method for class 'psp'
unmark(X)
 ## S3 method for class 'ppx'
unmark(X)

Arguments

X
A point pattern (object of class "ppp"), a split point pattern (object of class "splitppp"), a line segment pattern (object of class "psp") or a multidimensional space-time point pattern (object of class "ppx").

Details

A ‘mark’ is a value attached to each point in a spatial point pattern, or attached to each line segment in a line segment pattern, etc.

The function unmark is a simple way to remove the marks from such a dataset.

Value

An object of the same class as X with any mark information deleted.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

ppp.object, psp.object

Examples

  hicks <- lansing[lansing$marks == "hickory", ]

  # plot(hicks)  # still a marked point pattern, but only 1 value of marks
  # plot(unmark(hicks)) # unmarked

[Package spatstat.geom version 3.3-5 Index]