unstack.solist {spatstat.geom}	R Documentation

Unstack Each Spatial Object in a List of Objects

Description

Given a list of two-dimensional spatial objects, apply

Usage

## S3 method for class 'solist'
unstack(x, ...)

## S3 method for class 'layered'
unstack(x, ...)

Arguments

x
An object of class "solist" or "layered" representing a list of two-dimensional spatial objects.

...
Ignored.

Details

The functions defined here are methods for the generic unstack. They expect the argument x to be a list of spatial objects, of class "solist" or "layered".

Each spatial object in the list x will be unstacked by applying the relevant method for unstack. This means that

a marked point pattern with several columns of marks will be separated into several point patterns, each having a single column of marks

a measure with k-dimensional vector values will be separated into k measures with scalar values

The resulting unstacked objects will be collected into a list of the same kind as x. Typically the length of unstack(x) is greater than the length of x.

Value

A list belonging to the same class as x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

unstack

unstack.msr, unstack.ppp, unstack.psp

Examples

  A <- solist(finpines=finpines, cells=cells)
  A
  unstack(A)
  B <- layered(fin=finpines, loc=unmark(finpines),
              plotargs=list(list(), list(pch=16)))
  B
  plot(B)
  unstack(B)
  plot(unstack(B))

[Package spatstat.geom version 3.3-5 Index]