update.symbolmap {spatstat.geom}	R Documentation

Update a Graphics Symbol Map.

Description

This command updates the object using the arguments given.

Usage

## S3 method for class 'symbolmap'
update(object, ...)

Arguments

object
Graphics symbol map (object of class "symbolmap").

...
Additional or replacement arguments to symbolmap.

Details

This is a method for the generic function update for the class "symbolmap" of graphics symbol maps. It updates the object using the parameters given in the extra arguments ....

The extra arguments must be given in the form name=value and must be recognisable to symbolmap. They override any parameters of the same name in object.

Value

Another object of class "symbolmap".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

symbolmap to create a graphics symbol map.

Examples

  g <- symbolmap(size=function(x) x/50)
  g
  update(g, range=c(0,1))
  update(g, size=42)
  update(g, shape="squares", range=c(0,1))

[Package spatstat.geom version 3.3-5 Index]