venn.tess {spatstat.geom}	R Documentation

Tessellation Delimited by Several Sets

Description

Given a list of windows, construct the tessellation formed by all combinations of inclusion/exclusion of these windows.

Usage

venn.tess(..., window = NULL, labels=FALSE)

Arguments

...
Sets which delimit the tessellation. Any number of windows (objects of class "owin") or tessellations (objects of class "tess").

window
Optional. The bounding window of the resulting tessellation. If not specified, the default is the union of all the arguments ....

labels
Logical value, specifying whether to attach marks to each tile that reveal how it was formed.

Details

The arguments ... may be any number of windows. This function constructs a tessellation, like a Venn diagram, whose boundaries are made up of the boundaries of these sets. Each tile of the tessellation is defined by one of the possible combinations in which each set is either included or excluded.

If the arguments ... are named, then the resulting tiles will also have tile names, which identify the inclusion/exclusion combinations defining each tile. See the Examples.

If labels=TRUE then the tiles have marks which indicate the inclusion/exclusion combinations defining each tile. See the Examples.

Value

A tessellation (object of class "tess").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

intersect.tess.

To construct other kinds of tessellations, see tess, quadrats, hextess, polartess, dirichlet, delaunay, quantess and rpoislinetess.

Examples

   A <- square(1)
   B <- square(c(-0.5,0.5))
   W <- square(c(-1, 1.5))
   V <- venn.tess(A=A, B=B, window=W)
   V
   plot(V, do.labels=TRUE)
   Vlab <- venn.tess(A=A, B=B, window=W, labels=TRUE)
   marks(Vlab)

[Package spatstat.geom version 3.3-5 Index]