volume {spatstat.geom}	R Documentation

Volume of an Object

Description

Computes the volume of a spatial object such as a three-dimensional box.

Usage

 volume(x)

Arguments

x
An object whose volume will be computed.

Details

This function computes the volume of an object such as a three-dimensional box.

The function volume is generic, with methods for the classes "box3" (three-dimensional boxes) and "boxx" (multi-dimensional boxes).

There is also a method for the class "owin" (two-dimensional windows), which is identical to area.owin, and a method for the class "linnet" of linear networks, which returns the length of the network.

Value

The numerical value of the volume of the object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

area.owin, volume.box3, volume.boxx.

[Package spatstat.geom version 3.3-5 Index]