zapsmall.im {spatstat.geom}	R Documentation

Rounding of Pixel Values

Description

Modifies a pixel image, identifying those pixels that have values very close to zero, and replacing the value by zero.

Usage

  zapsmall.im(x, digits)

Arguments

x
Pixel image (object of class "im").

digits
Argument passed to zapsmall indicating the precision to be used.

Details

The function zapsmall is applied to each pixel value of the image x.

Value

Another pixel image.

Author(s)

Ege Rubak rubak@math.aau.dk and Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

zapsmall

Examples

  Z <- as.im(function(x,y) { exp(-40*(x+y)) }, square(1), dimyx=32)
  zapsmall.im(Z)

[Package spatstat.geom version 3.3-5 Index]