contour.imlist {spatstat.geom}	R Documentation

Array of Contour Plots

Description

Generates an array of contour plots.

Usage

  ## S3 method for class 'imlist'
contour(x, ...)

  ## S3 method for class 'listof'
contour(x, ...)

Arguments

x
An object of the class "imlist" representing a list of pixel images. Alternatively x may belong to the outdated class "listof".

...
Arguments passed to plot.solist to control the spatial arrangement of panels, and arguments passed to contour.im to control the display of each panel.

Details

This is a method for the generic command contour for the class "imlist". An object of class "imlist" represents a list of pixel images.

(The outdated class "listof" is also handled.)

Each entry in the list x will be displayed as a contour plot, in an array of panels laid out on the same graphics display, using plot.solist. Invididual panels are plotted by contour.im.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

plot.solist, contour.im

Examples

# bei.extra is a named list of covariate images
 contour(bei.extra,
         main="Barro Colorado: covariates")

[Package spatstat.geom version 3.3-5 Index]