coords {spatstat.geom}	R Documentation

Extract or Change Coordinates of a Spatial or Spatiotemporal Point Pattern

Description

Given any kind of spatial or space-time point pattern, this function extracts the (space and/or time and/or local) coordinates of the points and returns them as a data frame.

Usage

  coords(x, ...)
  ## S3 method for class 'ppp'
coords(x, ...)
  ## S3 method for class 'ppx'
coords(x, ..., spatial = TRUE, temporal = TRUE, local=TRUE)
  coords(x, ...) <- value
  ## S3 replacement method for class 'ppp'
coords(x, ...) <- value
  ## S3 replacement method for class 'ppx'
coords(x, ..., spatial = TRUE, temporal = TRUE, local=TRUE) <- value
  ## S3 method for class 'quad'
coords(x, ...)

Arguments

x
A point pattern: either a two-dimensional point pattern (object of class "ppp"), a three-dimensional point pattern (object of class "pp3"), or a general multidimensional space-time point pattern (object of class "ppx") or a quadrature scheme (object of class "quad").

...
Further arguments passed to methods.

spatial,temporal,local
Logical values indicating whether to extract spatial, temporal and local coordinates, respectively. The default is to return all such coordinates. (Only relevant to ppx objects).

value
New values of the coordinates. A numeric vector with one entry for each point in x, or a numeric matrix or data frame with one row for each point in x.

Details

The function coords extracts the coordinates from a point pattern. The function coords<- replaces the coordinates of the point pattern with new values.

Both functions coords and coords<- are generic, with methods for the classes "ppp") and "ppx". An object of class "pp3" also inherits from "ppx" and is handled by the method for "ppx".

Value

coords returns a data.frame with one row for each point, containing the coordinates. coords<- returns the altered point pattern.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

ppx, pp3, ppp, as.hyperframe.ppx, as.data.frame.ppx.

Examples

   df <- data.frame(x=runif(4),y=runif(4),t=runif(4))
   X <- ppx(data=df, coord.type=c("s","s","t"))
   coords(X)
   coords(X, temporal=FALSE)
   coords(X) <- matrix(runif(12), ncol=3)

[Package spatstat.geom version 3.3-5 Index]