default.dummy {spatstat.geom}	R Documentation

Generate a Default Pattern of Dummy Points

Description

Generates a default pattern of dummy points for use in a quadrature scheme.

Usage

 default.dummy(X, nd, random=FALSE, ntile=NULL, npix=NULL,
              quasi=FALSE, ..., eps=NULL, verbose=FALSE)

Arguments

X
The observed data point pattern. An object of class "ppp" or in a format recognised by as.ppp()

nd
Optional. Integer, or integer vector of length 2, specifying an nd * nd or nd[1] * nd[2] rectangular array of dummy points.

random
Logical value. If TRUE, the dummy points are generated randomly.

quasi
Logical value. If TRUE, the dummy points are generated by a quasirandom sequence.

ntile
Optional. Integer or pair of integers specifying the number of rows and columns of tiles used in the counting rule.

npix
Optional. Integer or pair of integers specifying the number of rows and columns of pixels used in computing approximate areas.

...
Ignored.

eps
Optional. Grid spacing. A positive number, or a vector of two positive numbers, giving the horizontal and vertical spacing, respectively, of the grid of dummy points. Incompatible with nd.

verbose
If TRUE, information about the construction of the quadrature scheme is printed.

Details

This function provides a sensible default for the dummy points in a quadrature scheme.

A quadrature scheme consists of the original data point pattern, an additional pattern of dummy points, and a vector of quadrature weights for all these points. See quad.object for further information about quadrature schemes.

If random and quasi are both false (the default), then the function creates dummy points in a regular nd[1] by nd[1] rectangular grid. If random is true and quasi is false, then the frame of the window is divided into an nd[1] by nd[1] array of tiles, and one dummy point is generated at random inside each tile. If quasi is true, a quasirandom pattern of nd[1] * nd[2] points is generated. In all cases, the four corner points of the frame of the window are added. Then if the window is not rectangular, any dummy points lying outside it are deleted.

If nd is missing, a default value is computed by the undocumented internal function default.n.tiling, using information about the data pattern X, and other arguments and settings. The default value of nd is always greater than or equal to spatstat.options("ndummy.min") and greater than or equal to 10 * ceiling(2 * sqrt(npoints(X))/10), and satisfies some other constraints. The default is designed so that model-fitting is relatively fast and stable, rather than highly accurate.

Alternative functions for creating dummy patterns include corners, gridcentres, stratrand and spokes.

Value

A point pattern (an object of class "ppp", see ppp.object) containing the dummy points.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

quad.object, quadscheme, corners, gridcentres, stratrand, spokes

Examples

  P <- simdat
  D <- default.dummy(P, 100)
  plot(D)
  Q <- quadscheme(P, D, "grid")
  if(interactive()) {plot(union.quad(Q))}

[Package spatstat.geom version 3.3-5 Index]