dilated.areas {spatstat.geom}	R Documentation

Areas of Morphological Dilations

Description

Computes the areas of successive morphological dilations.

Usage

  dilated.areas(X, r, W=as.owin(X), ..., constrained=TRUE, exact = FALSE)

Arguments

X
Object to be dilated. A point pattern (object of class "ppp"), a line segment pattern (object of class "psp"), or a window (object of class "owin").

r
Numeric vector of radii for the dilations.

W
Window (object of class "owin") inside which the areas will be computed, if constrained=TRUE.

...
Arguments passed to distmap to control the pixel resolution, if exact=FALSE.

constrained
Logical flag indicating whether areas should be restricted to the window W.

exact
Logical flag indicating whether areas should be computed using analytic geometry (which is slower but more accurate). Currently available only when X is a point pattern.

Details

This function computes the areas of the dilations of X by each of the radii r[i]. Areas may also be computed inside a specified window W.

The morphological dilation of a set X by a distance r > 0 is the subset consisting of all points x such that the distance from x to X is less than or equal to r.

When X is a point pattern, the dilation by a distance r is the union of discs of radius r centred at the points of X.

The argument r should be a vector of nonnegative numbers.

If exact=TRUE and if X is a point pattern, then the areas are computed using analytic geometry, which is slower but much more accurate. Otherwise the computation is performed using distmap.

To compute the dilated object itself, use dilation.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

owin, as.owin, dilation, eroded.areas

Examples

  X <- runifrect(10)
  a <- dilated.areas(X, c(0.1,0.2), W=square(1), exact=TRUE)

[Package spatstat.geom version 3.3-5 Index]