dirichletAreas {spatstat.geom}	R Documentation

Compute Areas of Tiles in Dirichlet Tessellation

Description

Calculates the area of each tile in the Dirichlet-Voronoi tessellation of a point pattern.

Usage

dirichletAreas(X)

Arguments

X
Point pattern (object of class "ppp").

Details

This is an efficient algorithm to calculate the areas of the tiles in the Dirichlet-Voronoi tessellation.

If the window of X is a binary pixel mask, the tile areas are computed by counting pixels. Otherwise the areas are computed exactly using analytic geometry.

If any points of X are duplicated, the duplicates will have tile area zero.

Value

Numeric vector with one entry for each point of X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

dirichlet, dirichletVertices

Examples

 aa <- dirichletAreas(cells)

[Package spatstat.geom version 3.3-5 Index]