edges {spatstat.geom}	R Documentation

Extract Boundary Edges of a Window.

Description

Extracts the boundary edges of a window and returns them as a line segment pattern.

Usage

  edges(x, ..., window = NULL, check = FALSE) 

Arguments

x
A window (object of class "owin"), or data acceptable to as.owin, specifying the window whose boundary is to be extracted.

...
Ignored.

window
Window to contain the resulting line segments. Defaults to as.rectangle(x).

check
Logical. Whether to check the validity of the resulting segment pattern.

Details

The boundary edges of the window x will be extracted as a line segment pattern.

Value

A line segment pattern (object of class "psp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

perimeter for calculating the total length of the boundary.

Examples

  edges(square(1))
  edges(letterR)

[Package spatstat.geom version 3.3-5 Index]