intensity {spatstat.geom}	R Documentation

Intensity of a Dataset or a Model

Description

Generic function for computing the intensity of a spatial dataset or spatial point process model.

Usage

intensity(X, ...)

Arguments

X
A spatial dataset or a spatial point process model.

...
Further arguments depending on the class of X.

Details

This is a generic function for computing the intensity of a spatial dataset or spatial point process model. There are methods for point patterns (objects of class "ppp") and fitted point process models (objects of class "ppm").

The empirical intensity of a dataset is the average density (the average amount of ‘stuff’ per unit area or volume). The empirical intensity of a point pattern is computed by the method intensity.ppp.

The theoretical intensity of a stochastic model is the expected density (expected amount of ‘stuff’ per unit area or volume). The theoretical intensity of a fitted point process model is computed by the method intensity.ppm.

Value

Usually a numeric value or vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

intensity.ppp, intensity.ppm.

[Package spatstat.geom version 3.3-5 Index]