methods.layered {spatstat.geom}	R Documentation

Methods for Layered Objects

Description

Methods for geometrical transformations of layered objects (class "layered").

Usage

  ## S3 method for class 'layered'
shift(X, vec=c(0,0), ...)

  ## S3 method for class 'layered'
rotate(X, ..., centre=NULL)

  ## S3 method for class 'layered'
affine(X, ...)

  ## S3 method for class 'layered'
reflect(X)

  ## S3 method for class 'layered'
flipxy(X)

  ## S3 method for class 'layered'
rescale(X, s, unitname)

  ## S3 method for class 'layered'
scalardilate(X, ...)

Arguments

X
Object of class "layered".

...
Arguments passed to the relevant methods when applying the operation to each layer of X.

s
Rescaling factor passed to the relevant method for rescale. May be missing.

vec
Shift vector (numeric vector of length 2).

centre
Centre of rotation. Either a vector of length 2, or a character string (partially matched to "centroid", "midpoint" or "bottomleft"). The default is the coordinate origin c(0,0).

unitname
Optional. New name for the unit of length. A value acceptable to the function unitname<-

Details

These are methods for the generic functions shift, rotate, reflect, affine, rescale, scalardilate and flipxy for the class of layered objects.

A layered object represents data that should be plotted in successive layers, for example, a background and a foreground. See layered.

Value

Another object of class "layered".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

layered

Examples

  B <- owin(c(5500, 9000), c(2500, 7400))
  L <- layered(Window(demopat), unmark(demopat)[B])
  plot(L)
  plot(rotate(L, pi/4))

[Package spatstat.geom version 3.3-5 Index]