pairdist.ppp {spatstat.geom}	R Documentation

Pairwise distances

Description

Computes the matrix of distances between all pairs of points in a point pattern.

Usage

  ## S3 method for class 'ppp'
pairdist(X, ...,
                        periodic=FALSE, method="C", squared=FALSE, metric=NULL)

Arguments

X
A point pattern (object of class "ppp").

...
Ignored.

periodic
Logical. Specifies whether to apply a periodic edge correction.

method
String specifying which method of calculation to use. Values are "C" and "interpreted". Usually not specified.

squared
Logical. If squared=TRUE, the squared distances are returned instead (this computation is faster).

metric
Optional. A metric (object of class "metric") that will be used to define and compute the distances.

Details

This is a method for the generic function pairdist.

Given a point pattern X (an object of class "ppp"), this function computes the Euclidean distances between all pairs of points in X, and returns the matrix of distances.

Alternatively if periodic=TRUE and the window containing X is a rectangle, then the distances will be computed in the ‘periodic’ sense (also known as ‘torus’ distance): opposite edges of the rectangle are regarded as equivalent. This is meaningless if the window is not a rectangle.

If squared=TRUE then the squared Euclidean distances d^2 are returned, instead of the Euclidean distances d. The squared distances are faster to calculate, and are sufficient for many purposes (such as finding the nearest neighbour of a point).

The argument method is not normally used. It is retained only for checking the validity of the software. If method = "interpreted" then the distances are computed using interpreted R code only. If method="C" (the default) then C code is used. The C code is somewhat faster.

Value

A square matrix whose [i,j] entry is the distance between the points numbered i and j.

Distance values

The values returned by pairdist(X) are distances, expressed as multiples of the unit of length of the spatial coordinates in X. The unit of length is given by unitname(X).

Note that, if the unit of length in X is a composite expression such as ‘2 microns’, then the values of pairdist(X) are expressed as multiples of 2 microns, rather than being expressed in microns.

Author(s)

Pavel Grabarnik pavel.grabar@issp.serpukhov.su and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

pairdist, pairdist.default, pairdist.psp, crossdist, nndist, Kest

Examples

   d <- pairdist(cells)
   d <- pairdist(cells, periodic=TRUE)
   d <- pairdist(cells, squared=TRUE)

[Package spatstat.geom version 3.3-5 Index]