quantile.im {spatstat.geom}	R Documentation

Sample Quantiles of Pixel Image

Description

Compute the sample quantiles of the pixel values of a given pixel image.

Usage

## S3 method for class 'im'
quantile(x, ...)

Arguments

x
A pixel image. An object of class "im".

...
Optional arguments passed to quantile.default. They determine the probabilities for which quantiles should be computed. See quantile.default.

Details

This simple function applies the generic quantile operation to the pixel values of the image x.

This function is a convenient way to inspect an image and to obtain summary statistics. See the examples.

Value

A vector of quantiles.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

quantile, cut.im, im.object

Examples

# artificial image data
Z <- setcov(square(1))

# find the quartiles
quantile(Z)

# find the deciles
quantile(Z, probs=(0:10)/10)

[Package spatstat.geom version 3.3-5 Index]