quantilefun.im {spatstat.geom}	R Documentation

Quantile Function for Images

Description

Return the inverse function of the cumulative distribution function of pixel values in an image.

Usage

  ## S3 method for class 'im'
quantilefun(x, ..., type=1)

Arguments

x
Pixel image (object of class "im").

...
Other arguments passed to methods.

type
Integer specifying the type of quantiles, as explained in quantile.default. Only types 1 and 2 are currently implemented.

Details

Whereas the command quantile calculates the quantiles of a dataset corresponding to desired probabilities p, the command quantilefun returns a function which can be used to compute any quantiles of the dataset.

If f <- quantilefun(x) then f is a function such that f(p) is the quantile associated with any given probability p. For example f(0.5) is the median of the original data, and f(0.99) is the 99th percentile of the original data.

If x is a pixel image (object of class "im") then the pixel values of x will be extracted and the quantile function of the pixel values is constructed.

Value

A function in the R language.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

quantilefun, ewcdf, quantile.ewcdf, ecdf, quantile

Examples

  ## image data: terrain elevation
  Z <- bei.extra$elev
  if(require(spatstat.explore)) {
    FE <- spatialcdf(Z, normalise=TRUE)
  } else {
    FE <- ecdf(Z[])
  }
  QE <- quantilefun(FE)
  QE(0.5) # median elevation
  if(interactive()) plot(QE, xlim=c(0,1),
     xlab="probability", ylab="quantile of elevation")

[Package spatstat.geom version 3.3-5 Index]