tilenames {spatstat.geom}	R Documentation

Names of Tiles in a Tessellation

Description

Extract or Change the Names of the Tiles in a Tessellation.

Usage

tilenames(x)
tilenames(x) <- value

## S3 method for class 'tess'
tilenames(x)

## S3 replacement method for class 'tess'
tilenames(x) <- value

Arguments

x
A tessellation (object of class "tess").

value
Character vector giving new names for the tiles.

Details

These functions extract or change the names of the tiles that make up the tessellation x.

If the tessellation is a regular grid, the tile names cannot be changed.

Value

tilenames returns a character vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

tess, tiles

Examples

  D <- dirichlet(runifrect(10))
  tilenames(D)
  tilenames(D) <- paste("Cell", 1:10)
  tilenames(D)

[Package spatstat.geom version 3.3-5 Index]