whichhalfplane {spatstat.geom}	R Documentation

Test Which Side of Infinite Line a Point Falls On

Description

Given an infinite line and a spatial point location, determine which side of the line the point falls on.

Usage

whichhalfplane(L, x, y = NULL)

Arguments

L
Object of class "infline" specifying one or more infinite straight lines in two dimensions.

x,y
Arguments acceptable to xy.coords specifying the locations of the points.

Details

An infinite line L divides the two-dimensional plane into two half-planes. This function returns a matrix M of logical values in which M[i,j] = TRUE if the jth spatial point lies below or to the left of the ith line.

Value

A logical matrix.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

infline

Examples

  L <- infline(p=runif(3), theta=runif(3, max=2*pi))
  X <- runifrect(4)
  whichhalfplane(L, X)

[Package spatstat.geom version 3.3-5 Index]